Rearranging code and removed "CS firstpass" ref.
- AudioFilters and Settings exiled
- removed CSharp firstpass ref (no more needed)
