+++ Chatterer/Chatterer.cs
@@ -138,14 +138,16 @@ namespace Chatterer
     //}
 
     [KSPAddon(KSPAddon.Startup.Flight, false)]
-    public class chatterer : MonoBehaviour
+    public partial class chatterer : MonoBehaviour
     {
+        //Version
+        private string this_version = "0.7.1.86";
+        private string main_window_title = "Chatterer ";
+        
         private static System.Random rand = new System.Random();
 
         private int window_base_id = -12381578;
-
-        private bool debugging = false;      //lots of extra log info if true
-
+        
         private Vessel vessel;          //is set to FlightGlobals.ActiveVessel
         private Vessel prev_vessel;     //to detect change in active vessel
 
@@ -207,7 +209,6 @@ namespace Chatterer
         private bool hide_all_windows = true;
         private string custom_dir_name = "directory name";  //default text for audioset input box
         private int active_menu = 0;    //selected main window section (sliders, sets, etc)
-        private int chatter_sel_filter;     //currently selected filter in filters window
         private int sel_beep_src = 0;   //currently selected beep source
         private int sel_beep_page = 1;
         private int num_beep_pages;
@@ -233,12 +234,7 @@ namespace Chatterer
 
         //Main window
         protected Rect main_window_pos = new Rect(Screen.width / 2f, Screen.height / 2f, 10f, 10f);
-
-        //Chatter filters window
-        private bool show_chatter_filter_settings = false;
-        protected Rect chatter_filter_settings_window_pos = new Rect(Screen.width / 2, Screen.height / 2, 10f, 10f);
-        private int chatter_filter_settings_window_id;
-
+        
         //Probe Sample Selector window
         private Vector2 probe_sample_selector_scroll_pos = new Vector2();
         private bool show_probe_sample_selector = false;
@@ -272,28 +268,7 @@ namespace Chatterer
         //private GUIStyle xkcd_label;
         private GUIStyle label_txt_bold;
         private GUIStyle button_txt_left_bold;
-
-        //Plugin settings
-        private bool run_once = true;   //used to run some things just once in Update() that don't get done in Awake()
-        //private bool power_available = true;
-        private bool quindar_toggle = true;
-        private bool disable_beeps_during_chatter = true;
-        private bool remotetech_toggle = false;
-        //private bool disable_power_usage = false;
-        private bool show_tooltips = true;
-        private bool http_update_check = false;
-        private bool use_vessel_settings = false;
-        private bool prev_use_vessel_settings = false;
-
-        //Chatter filters
-        private AudioChorusFilter chatter_chorus_filter;
-        private AudioDistortionFilter chatter_distortion_filter;
-        private AudioEchoFilter chatter_echo_filter;
-        private AudioHighPassFilter chatter_highpass_filter;
-        private AudioLowPassFilter chatter_lowpass_filter;
-        private AudioReverbFilter chatter_reverb_filter;
-        private int chatter_reverb_preset_index = 0;
-
+                        
         //Counters
         private float rt_update_timer = 0;
         private float sstv_timer = 0;
@@ -301,33 +276,7 @@ namespace Chatterer
         private float secs_since_last_exchange = 0;
         private float secs_since_initial_chatter = 0;
         private float secs_between_exchanges = 0;
-
-        //Sliders
-        private float chatter_freq_slider = 3f;
-        private int chatter_freq = 3;
-        private int prev_chatter_freq = 3;
-        private float chatter_vol_slider = 0.5f;
-        private float prev_chatter_vol_slider = 0.5f;
-
-        private float quindar_vol_slider = 0.5f;
-        private float prev_quindar_vol_slider = 0.5f;
-
-        private float sstv_freq_slider = 0;
-        private int sstv_freq = 0;
-        private int prev_sstv_freq = 0;
-        private float sstv_vol_slider = 0.25f;
-        private float prev_sstv_vol_slider = 0.25f;
-
-        //Insta-chatter key
-        private KeyCode insta_chatter_key = KeyCode.Slash;
-        private bool set_insta_chatter_key = false;
-        private bool insta_chatter_key_just_changed = false;
-
-        //Insta-SSTV key
-        private KeyCode insta_sstv_key = KeyCode.Quote;
-        private bool set_insta_sstv_key = false;
-        private bool insta_sstv_key_just_changed = false;
-
+        
         //these are updated and trigger chatter when changed
         private Vessel.Situations vessel_prev_sit;
         private int vessel_prev_stage;
@@ -355,46 +304,15 @@ namespace Chatterer
             controlDelay = 0; //delay from KSC
             //shortestcontrolDelay = 0; // delay from nearest sattelite
 
-        //Version
-        private string this_version = "0.7.1.86";
-        private string main_window_title = "Chatterer ";
-        
-        //Clipboards
-        private ConfigNode filters_clipboard;
-        private ConfigNode chorus_clipboard;
-        private ConfigNode dist_clipboard;
-        private ConfigNode echo_clipboard;
-        private ConfigNode hipass_clipboard;
-        private ConfigNode lopass_clipboard;
-        private ConfigNode reverb_clipboard;
-        private ConfigNode beepsource_clipboard;
-
-        //Settings nodes
-        private string settings_path;
-        private ConfigNode plugin_settings_node;
-        private ConfigNode vessel_settings_node;
-
         //Unsorted
         private BeepSource OTP_source = new BeepSource();
         private AudioClip OTP_stored_clip;    //holds the set probe sample while another sample plays once
         private bool OTP_playing = false;
-
-        private ConfigNode filter_defaults;
-
-        private bool mute_all = false;
-        private bool all_muted = false;
-
-        private bool show_advanced_options = false;
-        private bool show_chatter_sets = false;
-
-
+        
         private bool chatter_exists = false;
         private bool sstv_exists = false;
         private bool beeps_exists = false;
-
-
-        private string menu = "chatter";    //default to chatter menu because it has to have something
-
+        
         private List<GUISkin> g_skin_list;
 
         private string yep_yep = "";
@@ -1787,650 +1705,7 @@ namespace Chatterer
             GUILayout.EndVertical();
             GUI.DragWindow();
         }
-
-        private void chatter_filter_settings_gui(int window_id)
-        {
-            GUILayout.BeginVertical();
-
-            string[] filters = { "Chorus", "Dist", "Echo", "HiPass", "LoPass", "Reverb" };
-
-            chatter_sel_filter = GUILayout.SelectionGrid(chatter_sel_filter, filters, 3, GUILayout.ExpandWidth(true));
-
-            chatter_reverb_preset_index = combined_filter_settings_gui(chatter_sel_filter, chatter_chorus_filter, chatter_distortion_filter, chatter_echo_filter, chatter_highpass_filter, chatter_lowpass_filter, chatter_reverb_filter, chatter_reverb_preset_index);
-
-            GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-
-            GUIContent _content = new GUIContent();
-            _content.text = "Copy all";
-            _content.tooltip = "Copy all filter values to clipboard";
-            if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-            {
-                copy_all_chatter_filters();
-            }
-            if (filters_clipboard != null)
-            {
-                _content.text = "Paste all";
-                _content.tooltip = "Paste all filter values from clipboard";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    paste_all_chatter_filters();
-                }
-            }
-            GUILayout.Label(" ", GUILayout.ExpandWidth(true));  //spacer
-
-            if (GUILayout.Button("Close", GUILayout.ExpandWidth(false))) show_chatter_filter_settings = false;
-            GUILayout.EndHorizontal();
-
-            if (show_tooltips && GUI.tooltip != "") tooltips(chatter_filter_settings_window_pos);
-            //{
-            //    float w = 5.5f * GUI.tooltip.Length;
-            //    float x = (Event.current.mousePosition.x < chatter_filter_settings_window_pos.width / 2) ? Event.current.mousePosition.x + 10 : Event.current.mousePosition.x - 10 - w;
-            //    GUI.Box(new Rect(x, Event.current.mousePosition.y, w, 25f), GUI.tooltip, gs_tooltip);
-            //}
-
-            GUILayout.EndVertical();
-            GUI.DragWindow();
-        }
-
-        private void beep_filter_settings_gui(int window_id)
-        {
-            GUILayout.BeginVertical();
-
-            BeepSource source = null;
-
-            foreach (BeepSource bm in beepsource_list)
-            {
-                if (bm.settings_window_id == window_id)
-                {
-                    source = bm;
-                    break;
-                }
-            }
-
-            if (source != null)
-            {
-
-                string[] filters = { "Chorus", "Dist", "Echo", "HiPass", "LoPass", "Reverb" };
-
-                source.sel_filter = GUILayout.SelectionGrid(source.sel_filter, filters, 3, GUILayout.ExpandWidth(true));
-
-                source.reverb_preset_index = combined_filter_settings_gui(source.sel_filter, source.chorus_filter, source.distortion_filter, source.echo_filter, source.highpass_filter, source.lowpass_filter, source.reverb_filter, source.reverb_preset_index);
-
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-
-                GUIContent _content = new GUIContent();
-                _content.text = "Copy all";
-                _content.tooltip = "Copy all filter values to clipboard";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    copy_all_beep_filters(source);
-                }
-                if (filters_clipboard != null)
-                {
-                    _content.text = "Paste all";
-                    _content.tooltip = "Paste all filter values from clipboard";
-                    if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                    {
-                        paste_all_beep_filters(source);
-                    }
-                }
-                GUILayout.Label(" ", GUILayout.ExpandWidth(true));  //spacer
-
-                if (GUILayout.Button("Close", GUILayout.ExpandWidth(false))) source.show_settings_window = false;
-                GUILayout.EndHorizontal();
-
-                if (show_tooltips && GUI.tooltip != "") tooltips(source.settings_window_pos);
-                //{
-                //    float w = 5.5f * GUI.tooltip.Length;
-                //    float x = (Event.current.mousePosition.x < source.settings_window_pos.width / 2) ? Event.current.mousePosition.x + 10 : Event.current.mousePosition.x - 10 - w;
-                //    GUI.Box(new Rect(x, Event.current.mousePosition.y, w, 25f), GUI.tooltip, gs_tooltip);
-                //}
-
-
-                GUILayout.EndVertical();
-                GUI.DragWindow();
-            }
-        }
-
-        private int combined_filter_settings_gui(int sel_filter, AudioChorusFilter acf, AudioDistortionFilter adf, AudioEchoFilter aef, AudioHighPassFilter ahpf, AudioLowPassFilter alpf, AudioReverbFilter arf, int reverb_preset_index)
-        {
-            //chatter and beep settings window guis both call this function
-
-            GUIContent _content = new GUIContent();
-
-            if (sel_filter == 0)
-            {
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                _content.text = "Enable";
-                _content.tooltip = "Turn chorus filter on/off";
-                acf.enabled = GUILayout.Toggle(acf.enabled, _content, GUILayout.ExpandWidth(true));
-                _content.text = "Default";
-                _content.tooltip = "Reset chorus filter to default";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false))) reset_chorus_filter(acf);
-                GUILayout.EndHorizontal();
-
-                _content.text = "Dry mix: " + (acf.dryMix * 100).ToString("F0") + "%";
-                _content.tooltip = "Volume of original signal to pass to output";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                acf.dryMix = GUILayout.HorizontalSlider(acf.dryMix, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Wet mix 1: " + (acf.wetMix1 * 100).ToString("F0") + "%";
-                _content.tooltip = "Volume of 1st chorus tap";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                acf.wetMix1 = GUILayout.HorizontalSlider(acf.wetMix1, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Wet mix 2: " + (acf.wetMix2 * 100).ToString("F0") + "%";
-                _content.tooltip = "Volume of 2nd chorus tap";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                acf.wetMix2 = GUILayout.HorizontalSlider(acf.wetMix2, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Wet mix 3: " + (acf.wetMix3 * 100).ToString("F0") + "%";
-                _content.tooltip = "Volume of 3rd chorus tap";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                acf.wetMix3 = GUILayout.HorizontalSlider(acf.wetMix3, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Delay: " + acf.delay.ToString("F2") + " ms";
-                _content.tooltip = "Chorus delay in ms";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                acf.delay = GUILayout.HorizontalSlider(acf.delay, 0.1f, 100f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Rate: " + acf.rate.ToString("F2") + " Hz";
-                _content.tooltip = "Chorus modulation rate in Hz";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                acf.rate = GUILayout.HorizontalSlider(acf.rate, 0, 20f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Depth: " + (acf.depth * 100).ToString("F0") + "%";
-                _content.tooltip = "Chorus modulation depth";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                acf.depth = GUILayout.HorizontalSlider(acf.depth, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-                
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                _content.text = "Copy Chorus";
-                _content.tooltip = "Copy chorus values to clipboard";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    chorus_clipboard = new ConfigNode();
-                    chorus_clipboard.AddValue("dry_mix", acf.dryMix);
-                    chorus_clipboard.AddValue("wet_mix_1", acf.wetMix1);
-                    chorus_clipboard.AddValue("wet_mix_2", acf.wetMix2);
-                    chorus_clipboard.AddValue("wet_mix_3", acf.wetMix3);
-                    chorus_clipboard.AddValue("delay", acf.delay);
-                    chorus_clipboard.AddValue("rate", acf.rate);
-                    chorus_clipboard.AddValue("depth", acf.depth);
-                    
-                    if (debugging) Debug.Log("[CHATR] chorus filter values copied to chorus clipboard");
-                }
-                if (chorus_clipboard != null)
-                {
-                    _content.text = "Load Chorus";
-                    _content.tooltip = "Load chorus values from clipboard";
-                    if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                    {
-                        acf.dryMix = Single.Parse(chorus_clipboard.GetValue("dry_mix"));
-                        acf.wetMix1 = Single.Parse(chorus_clipboard.GetValue("wet_mix_1"));
-                        acf.wetMix2 = Single.Parse(chorus_clipboard.GetValue("wet_mix_2"));
-                        acf.wetMix3 = Single.Parse(chorus_clipboard.GetValue("wet_mix_3"));
-                        acf.delay = Single.Parse(chorus_clipboard.GetValue("delay"));
-                        acf.rate = Single.Parse(chorus_clipboard.GetValue("rate"));
-                        acf.depth = Single.Parse(chorus_clipboard.GetValue("depth"));
-                        
-                        if (debugging) Debug.Log("[CHATR] chorus filter values loaded from chorus clipboard");
-                    }
-                }
-                GUILayout.EndHorizontal();
-            }
-            else if (sel_filter == 1)
-            {
-                //Distortion
-                _content.text = "Enable";
-                _content.tooltip = "Turn distortion filter on/off";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                adf.enabled = GUILayout.Toggle(adf.enabled, _content, GUILayout.ExpandWidth(true));
-                _content.text = "Default";
-                _content.tooltip = "Reset distortion filter to default";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false))) reset_distortion_filter(adf);
-                GUILayout.EndHorizontal();
-
-                _content.text = "Distortion level: " + (adf.distortionLevel * 100).ToString("F0") + "%";
-                _content.tooltip = "Distortion value";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                adf.distortionLevel = GUILayout.HorizontalSlider(adf.distortionLevel, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                _content.text = "Copy Dist";
-                _content.tooltip = "Copy distortion values to clipboard";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    dist_clipboard = new ConfigNode();
-                    dist_clipboard.AddValue("distortion_level", adf.distortionLevel);
-                    if (debugging) Debug.Log("[CHATR] distortion filter values copied to distortion clipboard");
-                }
-                if (dist_clipboard != null)
-                {
-                    _content.text = "Load Dist";
-                    _content.tooltip = "Load distortion values from clipboard";
-                    if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                    {
-                        adf.distortionLevel = Single.Parse(dist_clipboard.GetValue("distortion_level"));
-                        if (debugging) Debug.Log("[CHATR] distortion filter values loaded from distortion clipboard");
-                    }
-                }
-                GUILayout.EndHorizontal();
-            }
-            else if (sel_filter == 2)
-            {
-                //Echo
-                _content.text = "Enable";
-                _content.tooltip = "Turn echo filter on/off";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                aef.enabled = GUILayout.Toggle(aef.enabled, _content, GUILayout.ExpandWidth(true));
-                _content.text = "Default";
-                _content.tooltip = "Reset echo filter to default";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false))) reset_echo_filter(aef);
-                GUILayout.EndHorizontal();
-
-                _content.text = "Delay: " + aef.delay.ToString("F0") + " ms";
-                _content.tooltip = "Echo delay in ms";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                aef.delay = GUILayout.HorizontalSlider(aef.delay, 10f, 5000f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Decay ratio: " + aef.decayRatio.ToString("F2");
-                _content.tooltip = "Echo decay per delay";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                aef.decayRatio = GUILayout.HorizontalSlider(aef.decayRatio, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Dry mix: " + (aef.dryMix * 100).ToString("F0") + "%";
-                _content.tooltip = "Volume of original signal to pass to output";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                aef.dryMix = GUILayout.HorizontalSlider(aef.dryMix, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Wet mix: " + (aef.wetMix * 100).ToString("F0") + "%";
-                _content.tooltip = "Volume of echo signal to pass to output";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                aef.wetMix = GUILayout.HorizontalSlider(aef.wetMix, 0, 1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                _content.text = "Copy Echo";
-                _content.tooltip = "Copy echo values to clipboard";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    echo_clipboard = new ConfigNode();
-                    echo_clipboard.AddValue("delay", aef.delay);
-                    echo_clipboard.AddValue("decay_ratio", aef.decayRatio);
-                    echo_clipboard.AddValue("dry_mix", aef.dryMix);
-                    echo_clipboard.AddValue("wet_mix", aef.wetMix);
-                    if (debugging) Debug.Log("[CHATR] echo filter values copied to echo clipboard");
-                }
-                if (echo_clipboard != null)
-                {
-                    _content.text = "Load Echo";
-                    _content.tooltip = "Load echo values from clipboard";
-                    if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                    {
-                        aef.delay = Single.Parse(echo_clipboard.GetValue("delay"));
-                        aef.decayRatio = Single.Parse(echo_clipboard.GetValue("decay_ratio"));
-                        aef.dryMix = Single.Parse(echo_clipboard.GetValue("dry_mix"));
-                        aef.wetMix = Single.Parse(echo_clipboard.GetValue("wet_mix"));
-                        if (debugging) Debug.Log("[CHATR] echo filter values loaded from echo clipboard");
-                    }
-                }
-                GUILayout.EndHorizontal();
-            }
-            else if (sel_filter == 3)
-            {
-                //Highpass
-                _content.text = "Enable";
-                _content.tooltip = "Turn highpass filter on/off";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                ahpf.enabled = GUILayout.Toggle(ahpf.enabled, _content, GUILayout.ExpandWidth(true));
-                _content.text = "Default";
-                _content.tooltip = "Reset highpass filter to default";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false))) reset_highpass_filter(ahpf);
-                GUILayout.EndHorizontal();
-
-                _content.text = "Cutoff freq: " + ahpf.cutoffFrequency.ToString("F2") + " Hz";
-                _content.tooltip = "Highpass cutoff frequency in Hz";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                ahpf.cutoffFrequency = GUILayout.HorizontalSlider(ahpf.cutoffFrequency, 10f, 22000f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Resonance Q: " + ahpf.highpassResonaceQ.ToString("F2");
-                _content.tooltip = "Highpass self-resonance dampening";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                ahpf.highpassResonaceQ = GUILayout.HorizontalSlider(ahpf.highpassResonaceQ, 1f, 10f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                _content.text = "Copy HiPass";
-                _content.tooltip = "Copy highpass values to clipboard";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    hipass_clipboard = new ConfigNode();
-                    hipass_clipboard.AddValue("cutoff_freq", ahpf.cutoffFrequency);
-                    hipass_clipboard.AddValue("resonance_q", ahpf.highpassResonaceQ);
-                    if (debugging) Debug.Log("[CHATR] highpass filter values copied to highpass clipboard");
-                }
-                if (hipass_clipboard != null)
-                {
-                    _content.text = "Load HiPass";
-                    _content.tooltip = "Load highpass values from clipboard";
-                    if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                    {
-                        ahpf.cutoffFrequency = Single.Parse(hipass_clipboard.GetValue("cutoff_freq"));
-                        ahpf.highpassResonaceQ = Single.Parse(hipass_clipboard.GetValue("resonance_q"));
-                        if (debugging) Debug.Log("[CHATR] highpass filter values loaded from highpass clipboard");
-                    }
-                }
-                GUILayout.EndHorizontal();
-            }
-            else if (sel_filter == 4)
-            {
-                //Lowpass
-                _content.text = "Enable";
-                _content.tooltip = "Turn lowpass filter on/off";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                alpf.enabled = GUILayout.Toggle(alpf.enabled, _content, GUILayout.ExpandWidth(true));
-                _content.text = "Default";
-                _content.tooltip = "Reset lowpass filter to default";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false))) reset_lowpass_filter(alpf);
-                GUILayout.EndHorizontal();
-
-                _content.text = "Cutoff freq: " + alpf.cutoffFrequency.ToString("F2") + " Hz";
-                _content.tooltip = "Lowpass cutoff frequency in Hz";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                alpf.cutoffFrequency = GUILayout.HorizontalSlider(alpf.cutoffFrequency, 10f, 22000f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Resonance Q: " + alpf.lowpassResonaceQ.ToString("F2");
-                _content.tooltip = "Lowpass self-resonance dampening";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                alpf.lowpassResonaceQ = GUILayout.HorizontalSlider(alpf.lowpassResonaceQ, 1f, 10f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                _content.text = "Copy LoPass";
-                _content.tooltip = "Copy lowpass values to clipboard";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    lopass_clipboard = new ConfigNode();
-                    lopass_clipboard.AddValue("cutoff_freq", alpf.cutoffFrequency);
-                    lopass_clipboard.AddValue("resonance_q", alpf.lowpassResonaceQ);
-                    if (debugging) Debug.Log("[CHATR] lowpass filter values copied to lowpass clipboard");
-                }
-                if (lopass_clipboard != null)
-                {
-                    _content.text = "Load LoPass";
-                    _content.tooltip = "Load lowpass values from clipboard";
-                    if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                    {
-                        alpf.cutoffFrequency = Single.Parse(lopass_clipboard.GetValue("cutoff_freq"));
-                        alpf.lowpassResonaceQ = Single.Parse(lopass_clipboard.GetValue("resonance_q"));
-                        if (debugging) Debug.Log("[CHATR] lowpass filter values loaded from lowpass clipboard");
-                    }
-                }
-                GUILayout.EndHorizontal();
-            }
-            else if (sel_filter == 5)
-            {
-                //Reverb
-                _content.text = "Enable";
-                _content.tooltip = "Turn reverb filter on/off";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                arf.enabled = GUILayout.Toggle(arf.enabled, _content, GUILayout.ExpandWidth(true));
-                _content.text = "Default";
-                _content.tooltip = "Reset reverb filter to default";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false))) reset_reverb_filter(arf);
-                GUILayout.EndHorizontal();
-
-                //Presets
-                AudioReverbPreset[] preset_array = Enum.GetValues(typeof(AudioReverbPreset)) as AudioReverbPreset[];
-                List<AudioReverbPreset> preset_list = new List<AudioReverbPreset>();
-
-                foreach (var val in preset_array)
-                {
-                    if (val.ToString() != "Off" && val.ToString() != "User") preset_list.Add(val);  //Off and User have separate buttons
-                }
-
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label("Current preset: " + arf.reverbPreset.ToString(), GUILayout.ExpandWidth(true));
-
-                //off button
-                //_content.text = "Off";
-                //_content.tooltip = "Turn reverb off";
-                //if (GUILayout.Button(_content, GUILayout.ExpandWidth(true)))
-                //{
-                //    arf.reverbPreset = AudioReverbPreset.Off;
-                //    if (debugging) Debug.Log("[CHATR] reverb turned off");
-                //}
-
-                //user button
-                _content.text = "User";
-                _content.tooltip = "User reverb settings";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    arf.reverbPreset = AudioReverbPreset.User;
-                    if (debugging) Debug.Log("[CHATR] reverb set to User");
-                }
-
-                GUILayout.EndHorizontal();
-
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-
-                if (GUILayout.Button("◄", GUILayout.ExpandWidth(true)))
-                {
-                    //need separate preset indices for chatter and each beep
-                    reverb_preset_index--;
-                    if (reverb_preset_index < 0) reverb_preset_index = preset_list.Count - 1;
-                    if (debugging) Debug.Log("[CHATR] reverb_preset_index = " + reverb_preset_index);
-                }
-                _content.text = preset_list[reverb_preset_index].ToString();
-                _content.tooltip = "Click to apply";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(true)))
-                {
-                    arf.reverbPreset = preset_list[reverb_preset_index];
-                    if (debugging) Debug.Log("[CHATR] reverb preset set to " + arf.reverbPreset.ToString());
-                }
-
-                if (GUILayout.Button("►", GUILayout.ExpandWidth(true)))
-                {
-                    reverb_preset_index++;
-                    if (reverb_preset_index >= preset_list.Count) reverb_preset_index = 0;
-                    if (debugging) Debug.Log("[CHATR] reverb_preset_index = " + reverb_preset_index);
-                }
-
-                GUILayout.EndHorizontal();
-
-                //GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                //GUILayout.Label("source.reverbPreset = " + arf.reverbPreset.ToString());
-                //GUILayout.EndHorizontal();
-
-                _content.text = "Dry level: " + arf.dryLevel.ToString("F0") + " mB";
-                _content.tooltip = "Mix level of dry signal in output in mB";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.dryLevel = GUILayout.HorizontalSlider(arf.dryLevel, -10000f, 0, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Room: " + arf.room.ToString("F0") + " mB";
-                _content.tooltip = "Room effect level at low frequencies in mB";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.room = GUILayout.HorizontalSlider(arf.room, -10000f, 0, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Room HF: " + arf.roomHF.ToString("F0") + " mB";
-                _content.tooltip = "Room effect high-frequency level in mB";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.roomHF = GUILayout.HorizontalSlider(arf.roomHF, -10000f, 0, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Room LF: " + arf.roomLF.ToString("F0") + " mB";
-                _content.tooltip = "Room effect low-frequency level in mB";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.roomLF = GUILayout.HorizontalSlider(arf.roomLF, -10000f, 0, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Room roll-off: " + arf.roomRolloff.ToString("F2");
-                _content.tooltip = "Rolloff factor for room effect";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.roomRolloff = GUILayout.HorizontalSlider(arf.roomRolloff, 0, 10f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Decay time: " + (arf.decayTime * 100).ToString("F0") + " s";
-                _content.tooltip = "Reverb decay time at low-frequencies";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.decayTime = GUILayout.HorizontalSlider(arf.decayTime, 0.1f, 20f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Decay HF ratio: " + arf.decayHFRatio.ToString("F0");
-                _content.tooltip = "HF to LF decay time ratio";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.decayHFRatio = GUILayout.HorizontalSlider(arf.decayHFRatio, 0.1f, 2f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Reflections level: " + arf.reflectionsLevel.ToString("F0") + " mB";
-                _content.tooltip = "Early reflections level relative to room effect";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.reflectionsLevel = GUILayout.HorizontalSlider(arf.reflectionsLevel, -10000f, 1000f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Reflections delay: " + arf.reflectionsDelay.ToString("F0") + " mB";
-                _content.tooltip = "Late reverberation level relative to room effect";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.reflectionsDelay = GUILayout.HorizontalSlider(arf.reflectionsDelay, -10000f, 2000f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Reverb level: " + arf.reverbLevel.ToString("F0") + " mB";
-                _content.tooltip = "Late reverberation level relative to room effect";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.reverbLevel = GUILayout.HorizontalSlider(arf.reverbLevel, -10000f, 2000f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Reverb delay: " + (arf.reverbDelay * 100).ToString("F0") + " s";
-                _content.tooltip = "Late reverb delay time rel. to first reflection";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.reverbDelay = GUILayout.HorizontalSlider(arf.reverbDelay, 0, 0.1f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Diffusion: " + arf.diffusion.ToString("F0") + "%";
-                _content.tooltip = "Reverb diffusion (echo density)";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.diffusion = GUILayout.HorizontalSlider(arf.diffusion, 0, 100f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "Density: " + arf.density.ToString("F0") + "%";
-                _content.tooltip = "Reverb density (modal density)";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.density = GUILayout.HorizontalSlider(arf.density, 0, 100f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "HF reference: " + arf.hfReference.ToString("F0") + " Hz";
-                _content.tooltip = "Reference high frequency in Hz";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.hfReference = GUILayout.HorizontalSlider(arf.hfReference, 20f, 20000f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                _content.text = "LF reference: " + arf.lFReference.ToString("F0") + " Hz";
-                _content.tooltip = "Reference low-frequency in Hz";
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                GUILayout.Label(_content, GUILayout.ExpandWidth(true));
-                arf.lFReference = GUILayout.HorizontalSlider(arf.lFReference, 20f, 1000f, GUILayout.Width(90f));
-                GUILayout.EndHorizontal();
-
-                GUILayout.BeginHorizontal(GUILayout.ExpandWidth(true));
-                _content.text = "Copy Reverb";
-                _content.tooltip = "Copy reverb values to clipboard";
-                if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                {
-                    reverb_clipboard = new ConfigNode();
-                    reverb_clipboard.AddValue("reverb_preset", arf.reverbPreset);
-                    reverb_clipboard.AddValue("dry_level", arf.dryLevel);
-                    reverb_clipboard.AddValue("room", arf.room);
-                    reverb_clipboard.AddValue("room_hf", arf.roomHF);
-                    reverb_clipboard.AddValue("room_lf", arf.roomLF);
-                    reverb_clipboard.AddValue("room_rolloff", arf.roomRolloff);
-                    reverb_clipboard.AddValue("decay_time", arf.decayTime);
-                    reverb_clipboard.AddValue("decay_hf_ratio", arf.decayHFRatio);
-                    reverb_clipboard.AddValue("reflections_level", arf.reflectionsLevel);
-                    reverb_clipboard.AddValue("reflections_delay", arf.reflectionsDelay);
-                    reverb_clipboard.AddValue("reverb_level", arf.reverbLevel);
-                    reverb_clipboard.AddValue("reverb_delay", arf.reverbDelay);
-                    reverb_clipboard.AddValue("diffusion", arf.diffusion);
-                    reverb_clipboard.AddValue("density", arf.density);
-                    reverb_clipboard.AddValue("hf_reference", arf.hfReference);
-                    reverb_clipboard.AddValue("lf_reference", arf.lFReference);
-                    if (debugging) Debug.Log("[CHATR] reverb filter values copied to reverb clipboard");
-                }
-                if (reverb_clipboard != null)
-                {
-                    _content.text = "Load Reverb";
-                    _content.tooltip = "Load reverb values from clipboard";
-                    if (GUILayout.Button(_content, GUILayout.ExpandWidth(false)))
-                    {
-                        arf.reverbPreset = (AudioReverbPreset)Enum.Parse(typeof(AudioReverbPreset), reverb_clipboard.GetValue("reverb_preset"));
-                        arf.dryLevel = Single.Parse(reverb_clipboard.GetValue("dry_level"));
-                        arf.room = Single.Parse(reverb_clipboard.GetValue("room"));
-                        arf.roomHF = Single.Parse(reverb_clipboard.GetValue("room_hf"));
-                        arf.roomLF = Single.Parse(reverb_clipboard.GetValue("room_lf"));
-                        arf.roomRolloff = Single.Parse(reverb_clipboard.GetValue("room_rolloff"));
-                        arf.decayTime = Single.Parse(reverb_clipboard.GetValue("decay_time"));
-                        arf.decayHFRatio = Single.Parse(reverb_clipboard.GetValue("decay_hf_ratio"));
-                        arf.reflectionsLevel = Single.Parse(reverb_clipboard.GetValue("reflections_level"));
-                        arf.reflectionsDelay = Single.Parse(reverb_clipboard.GetValue("reflections_delay"));
-                        arf.reverbLevel = Single.Parse(reverb_clipboard.GetValue("reverb_level"));
-                        arf.reverbDelay = Single.Parse(reverb_clipboard.GetValue("reverb_delay"));
-                        arf.diffusion = Single.Parse(reverb_clipboard.GetValue("diffusion"));
-                        arf.density = Single.Parse(reverb_clipboard.GetValue("density"));
-                        arf.hfReference = Single.Parse(reverb_clipboard.GetValue("hf_reference"));
-                        arf.lFReference = Single.Parse(reverb_clipboard.GetValue("lf_reference"));
-                        if (debugging) Debug.Log("[CHATR] reverb filter values loaded from reverb clipboard");
-                    }
-                }
-                GUILayout.EndHorizontal();
-            }
-            return reverb_preset_index;
-        }
-
+        
         private void probe_sample_selector_gui(int window_id)
         {
             GUIContent _content = new GUIContent();
@@ -2904,665 +2179,7 @@ namespace Chatterer
                 }
             }
         }
-
-        //Save/Load settings
-        private void save_plugin_settings()
-        {
-            //these values are not saved to vessel.cfg ever and are considered global
-            //if (debugging) Debug.Log("[CHATR] adding plugin settings to ConfigNode for write");
-            plugin_settings_node = new ConfigNode();
-            plugin_settings_node.name = "SETTINGS";
-            plugin_settings_node.AddValue("debugging", debugging);
-            plugin_settings_node.AddValue("use_vessel_settings", use_vessel_settings);
-            plugin_settings_node.AddValue("useBlizzy78Toolbar", useBlizzy78Toolbar);
-            plugin_settings_node.AddValue("http_update_check", http_update_check);
-            plugin_settings_node.AddValue("disable_beeps_during_chatter", disable_beeps_during_chatter);
-            plugin_settings_node.AddValue("insta_chatter_key", insta_chatter_key);
-            plugin_settings_node.AddValue("insta_sstv_key", insta_sstv_key);
-            plugin_settings_node.AddValue("show_advanced_options", show_advanced_options);
-
-            //also save values that are shared between the two configs
-            save_shared_settings(plugin_settings_node);
-
-            //save plugin.cfg
-            plugin_settings_node.Save(settings_path + "plugin.cfg");
-            if (debugging) Debug.Log("[CHATR] plugin settings saved to disk");
-
-            //update vessel_settings.cfg
-            if (use_vessel_settings)
-            {
-                write_vessel_settings();
-                if (debugging) Debug.Log("[CHATR] this vessel settings saved to disk");
-            }
-        }
-
-        private void load_plugin_settings()
-        {
-            if (debugging) Debug.Log("[CHATR] load_plugin_settings() START");
-
-            destroy_all_beep_players();
-            chatter_array.Clear();
-            beepsource_list.Clear();
-
-            plugin_settings_node = new ConfigNode();
-            plugin_settings_node = ConfigNode.Load(settings_path + "plugin.cfg");
-
-            if (plugin_settings_node != null)
-            {
-                if (debugging) Debug.Log("[CHATR] plugin_settings != null");
-                //Load settings specific to plugin.cfg
-                if (plugin_settings_node.HasValue("debugging")) debugging = Boolean.Parse(plugin_settings_node.GetValue("debugging"));
-                if (plugin_settings_node.HasValue("use_vessel_settings")) use_vessel_settings = Boolean.Parse(plugin_settings_node.GetValue("use_vessel_settings"));
-                if (plugin_settings_node.HasValue("useBlizzy78Toolbar")) useBlizzy78Toolbar = Boolean.Parse(plugin_settings_node.GetValue("useBlizzy78Toolbar"));
-                if (plugin_settings_node.HasValue("http_update_check")) http_update_check = Boolean.Parse(plugin_settings_node.GetValue("http_update_check"));
-                if (plugin_settings_node.HasValue("disable_beeps_during_chatter")) disable_beeps_during_chatter = Boolean.Parse(plugin_settings_node.GetValue("disable_beeps_during_chatter"));
-                if (plugin_settings_node.HasValue("insta_chatter_key")) insta_chatter_key = (KeyCode)Enum.Parse(typeof(KeyCode), plugin_settings_node.GetValue("insta_chatter_key"));
-                if (plugin_settings_node.HasValue("insta_sstv_key")) insta_sstv_key = (KeyCode)Enum.Parse(typeof(KeyCode), plugin_settings_node.GetValue("insta_sstv_key"));
-                if (plugin_settings_node.HasValue("show_advanced_options")) show_advanced_options = Boolean.Parse(plugin_settings_node.GetValue("show_advanced_options"));
-
-                load_shared_settings(plugin_settings_node); //load settings shared between both configs
-
-            }
-            else
-            {
-                if (debugging) Debug.LogWarning("[CHATR] plugin.cfg missing or unreadable");
-            }
-
-            //if (chatter_exists && chatter_array.Count == 0)
-            if (chatter_array.Count == 0)
-            {
-                if (debugging) Debug.Log("[CHATR] No audiosets found in config, adding defaults");
-                add_default_audiosets();
-                load_chatter_audio();   //load audio in case there is none
-            }
-
-            if (beeps_exists && beepsource_list.Count == 0)
-            {
-                if (debugging) Debug.LogWarning("[CHATR] beepsource_list.Count == 0, adding default 3");
-                add_default_beepsources();
-            }
-
-            if (aae_backgrounds_exist && backgroundsource_list.Count == 0)
-            {
-                if (debugging) Debug.LogWarning("[CHATR] backgroundsource_list.Count == 0, adding default 2");
-                add_default_backgroundsources();
-            }
-
-            if (debugging) Debug.Log("[CHATR] load_plugin_settings() END");
-        }
-
-        private void load_plugin_defaults()
-        {
-            if (debugging) Debug.Log("[CHATR] load_plugin_defaults() START");
-
-            destroy_all_beep_players();
-            chatter_array.Clear();
-            beepsource_list.Clear();
-
-            plugin_settings_node = new ConfigNode();
-            plugin_settings_node = ConfigNode.Load(settings_path + "plugin_defaults.cfg");
-
-            if (plugin_settings_node != null)
-            {
-                if (debugging) Debug.Log("[CHATR] plugin_defaults != null");
-                //Load settings specific to plugin.cfg
-                if (plugin_settings_node.HasValue("debugging")) debugging = Boolean.Parse(plugin_settings_node.GetValue("debugging"));
-                if (plugin_settings_node.HasValue("use_vessel_settings")) use_vessel_settings = Boolean.Parse(plugin_settings_node.GetValue("use_vessel_settings"));
-                if (plugin_settings_node.HasValue("useBlizzy78Toolbar")) useBlizzy78Toolbar = Boolean.Parse(plugin_settings_node.GetValue("useBlizzy78Toolbar"));
-                if (plugin_settings_node.HasValue("http_update_check")) http_update_check = Boolean.Parse(plugin_settings_node.GetValue("http_update_check"));
-                if (plugin_settings_node.HasValue("disable_beeps_during_chatter")) disable_beeps_during_chatter = Boolean.Parse(plugin_settings_node.GetValue("disable_beeps_during_chatter"));
-                if (plugin_settings_node.HasValue("insta_chatter_key")) insta_chatter_key = (KeyCode)Enum.Parse(typeof(KeyCode), plugin_settings_node.GetValue("insta_chatter_key"));
-                if (plugin_settings_node.HasValue("insta_sstv_key")) insta_sstv_key = (KeyCode)Enum.Parse(typeof(KeyCode), plugin_settings_node.GetValue("insta_sstv_key"));
-                if (plugin_settings_node.HasValue("show_advanced_options")) show_advanced_options = Boolean.Parse(plugin_settings_node.GetValue("show_advanced_options"));
-
-                load_shared_settings(plugin_settings_node); //load settings shared between both configs
-
-            }
-            else
-            {
-                if (debugging) Debug.LogWarning("[CHATR] plugin_defautls.cfg missing or unreadable");
-            }
-
-            //if (chatter_exists && chatter_array.Count == 0)
-            if (chatter_array.Count == 0)
-            {
-                if (debugging) Debug.Log("[CHATR] No audiosets found in config, adding defaults");
-                add_default_audiosets();
-                load_chatter_audio();   //load audio in case there is none
-            }
-
-            if (beeps_exists && beepsource_list.Count == 0)
-            {
-                if (debugging) Debug.LogWarning("[CHATR] beepsource_list.Count == 0, adding default 3");
-                add_default_beepsources();
-            }
-
-            if (aae_backgrounds_exist && backgroundsource_list.Count == 0)
-            {
-                if (debugging) Debug.LogWarning("[CHATR] backgroundsource_list.Count == 0, adding default 2");
-                add_default_backgroundsources();
-            }
-
-            if (debugging) Debug.Log("[CHATR] load_plugin_defaults() END");
-        }
-
-        //Functions to handle settings shared by plugin.cfg and vessel.cfg
-        private void save_shared_settings(ConfigNode node)
-        {
-            node.AddValue("show_tooltips", show_tooltips);
-            node.AddValue("main_window_pos", main_window_pos.x + "," + main_window_pos.y);
-            node.AddValue("hide_all_windows", hide_all_windows);
-            node.AddValue("skin_index", skin_index);
-            node.AddValue("active_menu", active_menu);
-            node.AddValue("remotetech_toggle", remotetech_toggle);
-
-            node.AddValue("chatter_freq", chatter_freq);
-            node.AddValue("chatter_vol_slider", chatter_vol_slider);
-            node.AddValue("chatter_sel_filter", chatter_sel_filter);
-            node.AddValue("show_chatter_filter_settings", show_chatter_filter_settings);
-            node.AddValue("show_sample_selector", show_probe_sample_selector);
-            node.AddValue("chatter_reverb_preset_index", chatter_reverb_preset_index);
-            node.AddValue("chatter_filter_settings_window_pos", chatter_filter_settings_window_pos.x + "," + chatter_filter_settings_window_pos.y);
-            node.AddValue("probe_sample_selector_window_pos", probe_sample_selector_window_pos.x + "," + probe_sample_selector_window_pos.y);
-
-            node.AddValue("quindar_toggle", quindar_toggle);
-            node.AddValue("quindar_vol_slider", quindar_vol_slider);
-            node.AddValue("sstv_freq", sstv_freq);
-            node.AddValue("sstv_vol_slider", sstv_vol_slider);
-
-            node.AddValue("sel_beep_src", sel_beep_src);
-            node.AddValue("sel_beep_page", sel_beep_page);
-
-            //AAE
-            if (aae_backgrounds_exist)
-            {
-                foreach (BackgroundSource src in backgroundsource_list)
-                {
-                    ConfigNode _background = new ConfigNode();
-                    _background.name = "AAE_BACKGROUND";
-                    _background.AddValue("volume", src.audiosource.volume);
-                    _background.AddValue("current_clip", src.current_clip);
-                    node.AddNode(_background);
-                }
-            }
-
-            if (aae_soundscapes_exist)
-            {
-                node.AddValue("aae_soundscape_vol", aae_soundscape.volume);
-                node.AddValue("aae_soundscape_freq", aae_soundscape_freq);
-            }
-
-            if (aae_breathing_exist) node.AddValue("aae_breathing_vol", aae_breathing.volume);
-            if (aae_wind_exist) node.AddValue("aae_wind_vol", aae_wind_vol_slider);
-            if (aae_airlock_exist) node.AddValue("aae_airlock_vol", aae_airlock.volume);
-
-            //Chatter sets
-            foreach (ChatterAudioList chatter_set in chatter_array)
-            {
-                ConfigNode _set = new ConfigNode();
-                _set.name = "AUDIOSET";
-                _set.AddValue("directory", chatter_set.directory);
-                _set.AddValue("is_active", chatter_set.is_active);
-                node.AddNode(_set);
-            }
-
-            //filters
-            ConfigNode _filter;
-            _filter = new ConfigNode();
-            _filter.name = "CHORUS";
-            _filter.AddValue("enabled", chatter_chorus_filter.enabled);
-            _filter.AddValue("dry_mix", chatter_chorus_filter.dryMix);
-            _filter.AddValue("wet_mix_1", chatter_chorus_filter.wetMix1);
-            _filter.AddValue("wet_mix_2", chatter_chorus_filter.wetMix2);
-            _filter.AddValue("wet_mix_3", chatter_chorus_filter.wetMix3);
-            _filter.AddValue("delay", chatter_chorus_filter.delay);
-            _filter.AddValue("rate", chatter_chorus_filter.rate);
-            _filter.AddValue("depth", chatter_chorus_filter.depth);
-            node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "DISTORTION";
-            _filter.AddValue("enabled", chatter_distortion_filter.enabled);
-            _filter.AddValue("distortion_level", chatter_distortion_filter.distortionLevel);
-            node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "ECHO";
-            _filter.AddValue("enabled", chatter_echo_filter.enabled);
-            _filter.AddValue("delay", chatter_echo_filter.delay);
-            _filter.AddValue("decay_ratio", chatter_echo_filter.decayRatio);
-            _filter.AddValue("dry_mix", chatter_echo_filter.dryMix);
-            _filter.AddValue("wet_mix", chatter_echo_filter.wetMix);
-            node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "HIGHPASS";
-            _filter.AddValue("enabled", chatter_highpass_filter.enabled);
-            _filter.AddValue("cutoff_freq", chatter_highpass_filter.cutoffFrequency);
-            _filter.AddValue("resonance_q", chatter_highpass_filter.highpassResonaceQ);
-            node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "LOWPASS";
-            _filter.AddValue("enabled", chatter_lowpass_filter.enabled);
-            _filter.AddValue("cutoff_freq", chatter_lowpass_filter.cutoffFrequency);
-            _filter.AddValue("resonance_q", chatter_lowpass_filter.lowpassResonaceQ);
-            node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "REVERB";
-            _filter.AddValue("enabled", chatter_reverb_filter.enabled);
-            _filter.AddValue("reverb_preset", chatter_reverb_filter.reverbPreset);
-            _filter.AddValue("dry_level", chatter_reverb_filter.dryLevel);
-            _filter.AddValue("room", chatter_reverb_filter.room);
-            _filter.AddValue("room_hf", chatter_reverb_filter.roomHF);
-            _filter.AddValue("room_lf", chatter_reverb_filter.roomLF);
-            _filter.AddValue("room_rolloff", chatter_reverb_filter.roomRolloff);
-            _filter.AddValue("decay_time", chatter_reverb_filter.decayTime);
-            _filter.AddValue("decay_hf_ratio", chatter_reverb_filter.decayHFRatio);
-            _filter.AddValue("reflections_level", chatter_reverb_filter.reflectionsLevel);
-            _filter.AddValue("reflections_delay", chatter_reverb_filter.reflectionsDelay);
-            _filter.AddValue("reverb_level", chatter_reverb_filter.reverbLevel);
-            _filter.AddValue("reverb_delay", chatter_reverb_filter.reverbDelay);
-            _filter.AddValue("diffusion", chatter_reverb_filter.diffusion);
-            _filter.AddValue("density", chatter_reverb_filter.density);
-            _filter.AddValue("hf_reference", chatter_reverb_filter.hfReference);
-            _filter.AddValue("lf_reference", chatter_reverb_filter.lFReference);
-            node.AddNode(_filter);
-
-
-            foreach (BeepSource source in beepsource_list)
-            {
-                ConfigNode beep_settings = new ConfigNode();
-                beep_settings.name = "BEEPSOURCE";
-
-                beep_settings.AddValue("precise", source.precise);
-                beep_settings.AddValue("precise_freq", source.precise_freq);
-                beep_settings.AddValue("loose_freq", source.loose_freq);
-                beep_settings.AddValue("volume", source.audiosource.volume);
-                beep_settings.AddValue("pitch", source.audiosource.pitch);
-                beep_settings.AddValue("current_clip", source.current_clip);
-                beep_settings.AddValue("sel_filter", source.sel_filter);
-                beep_settings.AddValue("show_settings_window", source.show_settings_window);
-                beep_settings.AddValue("reverb_preset_index", source.reverb_preset_index);
-                beep_settings.AddValue("settings_window_pos_x", source.settings_window_pos.x);
-                beep_settings.AddValue("settings_window_pos_y", source.settings_window_pos.y);
-
-                //filters
-                //ConfigNode _filter;
-
-                _filter = new ConfigNode();
-                _filter.name = "CHORUS";
-                _filter.AddValue("enabled", source.chorus_filter.enabled);
-                _filter.AddValue("dry_mix", source.chorus_filter.dryMix);
-                _filter.AddValue("wet_mix_1", source.chorus_filter.wetMix1);
-                _filter.AddValue("wet_mix_2", source.chorus_filter.wetMix2);
-                _filter.AddValue("wet_mix_3", source.chorus_filter.wetMix3);
-                _filter.AddValue("delay", source.chorus_filter.delay);
-                _filter.AddValue("rate", source.chorus_filter.rate);
-                _filter.AddValue("depth", source.chorus_filter.depth);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "DISTORTION";
-                _filter.AddValue("enabled", source.distortion_filter.enabled);
-                _filter.AddValue("distortion_level", source.distortion_filter.distortionLevel);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "ECHO";
-                _filter.AddValue("enabled", source.echo_filter.enabled);
-                _filter.AddValue("delay", source.echo_filter.delay);
-                _filter.AddValue("decay_ratio", source.echo_filter.decayRatio);
-                _filter.AddValue("dry_mix", source.echo_filter.dryMix);
-                _filter.AddValue("wet_mix", source.echo_filter.wetMix);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "HIGHPASS";
-                _filter.AddValue("enabled", source.highpass_filter.enabled);
-                _filter.AddValue("cutoff_freq", source.highpass_filter.cutoffFrequency);
-                _filter.AddValue("resonance_q", source.highpass_filter.highpassResonaceQ);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "LOWPASS";
-                _filter.AddValue("enabled", source.lowpass_filter.enabled);
-                _filter.AddValue("cutoff_freq", source.lowpass_filter.cutoffFrequency);
-                _filter.AddValue("resonance_q", source.lowpass_filter.lowpassResonaceQ);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "REVERB";
-                _filter.AddValue("enabled", source.reverb_filter.enabled);
-                _filter.AddValue("reverb_preset", source.reverb_filter.reverbPreset);
-                _filter.AddValue("dry_level", source.reverb_filter.dryLevel);
-                _filter.AddValue("room", source.reverb_filter.room);
-                _filter.AddValue("room_hf", source.reverb_filter.roomHF);
-                _filter.AddValue("room_lf", source.reverb_filter.roomLF);
-                _filter.AddValue("room_rolloff", source.reverb_filter.roomRolloff);
-                _filter.AddValue("decay_time", source.reverb_filter.decayTime);
-                _filter.AddValue("decay_hf_ratio", source.reverb_filter.decayHFRatio);
-                _filter.AddValue("reflections_level", source.reverb_filter.reflectionsLevel);
-                _filter.AddValue("reflections_delay", source.reverb_filter.reflectionsDelay);
-                _filter.AddValue("reverb_level", source.reverb_filter.reverbLevel);
-                _filter.AddValue("reverb_delay", source.reverb_filter.reverbDelay);
-                _filter.AddValue("diffusion", source.reverb_filter.diffusion);
-                _filter.AddValue("density", source.reverb_filter.density);
-                _filter.AddValue("hf_reference", source.reverb_filter.hfReference);
-                _filter.AddValue("lf_reference", source.reverb_filter.lFReference);
-                beep_settings.AddNode(_filter);
-
-                node.AddNode(beep_settings);
-            }
-        }
-
-        private void load_shared_settings(ConfigNode node)
-        {
-            if (debugging) Debug.Log("[CHATR] load_shared_settings() START");
-
-            destroy_all_beep_players();
-            beepsource_list.Clear();
-            destroy_all_background_players();
-            backgroundsource_list.Clear();
-            chatter_array.Clear();
-
-            if (node.HasValue("main_window_pos"))
-            {
-                string[] split = node.GetValue("main_window_pos").Split(Convert.ToChar(","));
-                main_window_pos.x = Single.Parse(split[0]);
-                main_window_pos.y = Single.Parse(split[1]);
-            }
-            
-            if (node.HasValue("show_tooltips")) show_tooltips = Boolean.Parse(node.GetValue("show_tooltips"));
-            if (node.HasValue("hide_all_windows")) hide_all_windows = Boolean.Parse(node.GetValue("hide_all_windows"));
-            if (node.HasValue("skin_index")) skin_index = Int32.Parse(node.GetValue("skin_index"));
-            if (node.HasValue("active_menu")) active_menu = Int32.Parse(node.GetValue("active_menu"));
-            if (node.HasValue("remotetech_toggle")) remotetech_toggle = Boolean.Parse(node.GetValue("remotetech_toggle"));
-
-            if (node.HasValue("chatter_freq"))
-            {
-                chatter_freq = Int32.Parse(node.GetValue("chatter_freq"));
-                chatter_freq_slider = chatter_freq;
-                prev_chatter_freq = chatter_freq;
-            }
-            if (node.HasValue("chatter_vol_slider"))
-            {
-                chatter_vol_slider = Single.Parse(node.GetValue("chatter_vol_slider"));
-                initial_chatter.volume = chatter_vol_slider;
-                response_chatter.volume = chatter_vol_slider;
-                prev_chatter_vol_slider = chatter_vol_slider;
-            }
-            if (node.HasValue("chatter_sel_filter")) chatter_sel_filter = Int32.Parse(node.GetValue("chatter_sel_filter"));
-            if (node.HasValue("show_chatter_filter_settings")) show_chatter_filter_settings = Boolean.Parse(node.GetValue("show_chatter_filter_settings"));
-            if (node.HasValue("chatter_reverb_preset_index")) chatter_reverb_preset_index = Int32.Parse(node.GetValue("chatter_reverb_preset_index"));
-            if (node.HasValue("chatter_filter_settings_window_pos"))
-            {
-                string[] split = node.GetValue("chatter_filter_settings_window_pos").Split(Convert.ToChar(","));
-                chatter_filter_settings_window_pos.x = Single.Parse(split[0]);
-                chatter_filter_settings_window_pos.y = Single.Parse(split[1]);
-            }
-            if (node.HasValue("show_sample_selector")) show_probe_sample_selector = Boolean.Parse(node.GetValue("show_sample_selector"));
-            if (node.HasValue("probe_sample_selector_window_pos"))
-            {
-                string[] split = node.GetValue("probe_sample_selector_window_pos").Split(Convert.ToChar(","));
-                probe_sample_selector_window_pos.x = Single.Parse(split[0]);
-                probe_sample_selector_window_pos.y = Single.Parse(split[1]);
-            }
-            if (node.HasValue("quindar_toggle")) quindar_toggle = Boolean.Parse(node.GetValue("quindar_toggle"));
-            if (node.HasValue("quindar_vol_slider"))
-            {
-                quindar_vol_slider = Single.Parse(node.GetValue("quindar_vol_slider"));
-                prev_quindar_vol_slider = quindar_vol_slider;
-            }
-            if (node.HasValue("sstv_freq"))
-            {
-                sstv_freq = Int32.Parse(node.GetValue("sstv_freq"));
-                sstv_freq_slider = sstv_freq;
-                prev_sstv_freq = sstv_freq;
-            }
-            if (node.HasValue("sstv_vol_slider"))
-            {
-                sstv_vol_slider = Single.Parse(node.GetValue("sstv_vol_slider"));
-                prev_sstv_vol_slider = sstv_vol_slider;
-            }
-            if (node.HasValue("sel_beep_src")) sel_beep_src = Int32.Parse(node.GetValue("sel_beep_src"));
-            if (sel_beep_src < 0 || sel_beep_src > 9) sel_beep_src = 0;
-            if (node.HasValue("sel_beep_page")) sel_beep_page = Int32.Parse(node.GetValue("sel_beep_page"));
-
-            //AAE
-            int i;
-
-            if (aae_backgrounds_exist)
-            {
-                i = 0;
-
-                foreach (ConfigNode _background in node.nodes)
-                {
-                    if (_background.name == "AAE_BACKGROUND")
-                    {
-                        add_new_backgroundsource();
-                        if (_background.HasValue("volume")) backgroundsource_list[i].audiosource.volume = Single.Parse(_background.GetValue("volume"));
-                        if (_background.HasValue("current_clip")) backgroundsource_list[i].current_clip = _background.GetValue("current_clip");
-
-                        if (dict_background_samples.Count > 0)
-                        {
-                            set_background_clip(backgroundsource_list[i]);
-                        }
-                        i++;
-                    }
-                }
-            }
-
-            if (aae_soundscapes_exist)
-            {
-                if (node.HasValue("aae_soundscape_vol")) aae_soundscape.volume = Single.Parse(node.GetValue("aae_soundscape_vol"));
-                if (node.HasValue("aae_soundscape_freq"))
-                {
-                    aae_soundscape_freq = Int32.Parse(node.GetValue("aae_soundscape_freq"));
-                    aae_prev_soundscape_freq = aae_soundscape_freq;
-                }
-            }
-
-            if (aae_breathing_exist)
-            {
-                if (node.HasValue("aae_breathing_vol")) aae_breathing.volume = Single.Parse(node.GetValue("aae_breathing_vol"));
-            }
-
-            if (aae_airlock_exist)
-            {
-                if (node.HasValue("aae_airlock_vol")) aae_airlock.volume = Single.Parse(node.GetValue("aae_airlock_vol"));
-            }
-
-            if (aae_wind_exist)
-            {
-                if (node.HasValue("aae_wind_vol"))
-                {
-                    aae_wind_vol_slider = Single.Parse(node.GetValue("aae_wind_vol"));
-                    aae_wind.volume = aae_wind_vol_slider;
-                }
-            }
-
-            //
-            //Load audioset info
-            i = 0;
-            foreach (ConfigNode _set in node.nodes)
-            {
-                if (_set.name == "AUDIOSET")
-                {
-                    chatter_array.Add(new ChatterAudioList());  //create a new entry in the list for each audioset
-                    if (_set.HasValue("directory")) chatter_array[i].directory = _set.GetValue("directory");
-                    if (_set.HasValue("is_active")) chatter_array[i].is_active = Boolean.Parse(_set.GetValue("is_active"));
-                    i++;
-                }
-            }
-            if (debugging) Debug.Log("[CHATR] audiosets found: " + chatter_array.Count + " :: reloading chatter audio");
-            load_chatter_audio();   //reload audio
-
-            //Chatter filters
-            foreach (ConfigNode _filter in node.nodes)
-            {
-                if (_filter.name == "CHORUS")
-                {
-                    if (_filter.HasValue("enabled"))
-                    {
-                        chatter_chorus_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                        //if (debugging) Debug.Log("[CHATR] beepsource index: " + x.ToString() + " :: chorus_enabled = " + beepsource_list[x].chorus_filter.enabled);
-                    }
-                    if (_filter.HasValue("dry_mix")) chatter_chorus_filter.dryMix = Single.Parse(_filter.GetValue("dry_mix"));
-                    if (_filter.HasValue("wet_mix_1")) chatter_chorus_filter.wetMix1 = Single.Parse(_filter.GetValue("wet_mix_1"));
-                    if (_filter.HasValue("wet_mix_2")) chatter_chorus_filter.wetMix2 = Single.Parse(_filter.GetValue("wet_mix_2"));
-                    if (_filter.HasValue("wet_mix_3")) chatter_chorus_filter.wetMix3 = Single.Parse(_filter.GetValue("wet_mix_3"));
-                }
-                else if (_filter.name == "DISTORTION")
-                {
-                    if (_filter.HasValue("enabled")) chatter_distortion_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                    if (_filter.HasValue("distortion_level")) chatter_distortion_filter.distortionLevel = Single.Parse(_filter.GetValue("distortion_level"));
-                }
-                else if (_filter.name == "ECHO")
-                {
-                    if (_filter.HasValue("enabled")) chatter_echo_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                    if (_filter.HasValue("delay")) chatter_echo_filter.delay = Single.Parse(_filter.GetValue("delay"));
-                    if (_filter.HasValue("decay_ratio")) chatter_echo_filter.decayRatio = Single.Parse(_filter.GetValue("decay_ratio"));
-                    if (_filter.HasValue("dry_mix")) chatter_echo_filter.dryMix = Single.Parse(_filter.GetValue("dry_mix"));
-                    if (_filter.HasValue("wet_mix")) chatter_echo_filter.wetMix = Single.Parse(_filter.GetValue("wet_mix"));
-                }
-                else if (_filter.name == "HIGHPASS")
-                {
-                    if (_filter.HasValue("enabled")) chatter_highpass_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                    if (_filter.HasValue("cutoff_freq")) chatter_highpass_filter.cutoffFrequency = Single.Parse(_filter.GetValue("cutoff_freq"));
-                    if (_filter.HasValue("resonance_q")) chatter_highpass_filter.highpassResonaceQ = Single.Parse(_filter.GetValue("resonance_q"));
-                }
-                else if (_filter.name == "LOWPASS")
-                {
-                    if (_filter.HasValue("enabled")) chatter_lowpass_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                    if (_filter.HasValue("cutoff_freq")) chatter_lowpass_filter.cutoffFrequency = Single.Parse(_filter.GetValue("cutoff_freq"));
-                    if (_filter.HasValue("resonance_q")) chatter_lowpass_filter.lowpassResonaceQ = Single.Parse(_filter.GetValue("resonance_q"));
-                }
-                else if (_filter.name == "REVERB")
-                {
-                    if (_filter.HasValue("enabled")) chatter_reverb_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                    if (_filter.HasValue("reverb_preset")) chatter_reverb_filter.reverbPreset = (AudioReverbPreset)Enum.Parse(typeof(AudioReverbPreset), _filter.GetValue("reverb_preset"));
-                    if (_filter.HasValue("dry_level")) chatter_reverb_filter.dryLevel = Single.Parse(_filter.GetValue("dry_level"));
-                    if (_filter.HasValue("room")) chatter_reverb_filter.room = Single.Parse(_filter.GetValue("room"));
-                    if (_filter.HasValue("room_hf")) chatter_reverb_filter.roomHF = Single.Parse(_filter.GetValue("room_hf"));
-                    if (_filter.HasValue("room_lf")) chatter_reverb_filter.roomLF = Single.Parse(_filter.GetValue("room_lf"));
-                    if (_filter.HasValue("room_rolloff")) chatter_reverb_filter.roomRolloff = Single.Parse(_filter.GetValue("room_rolloff"));
-                    if (_filter.HasValue("decay_time")) chatter_reverb_filter.decayTime = Single.Parse(_filter.GetValue("decay_time"));
-                    if (_filter.HasValue("decay_hf_ratio")) chatter_reverb_filter.decayHFRatio = Single.Parse(_filter.GetValue("decay_hf_ratio"));
-                    if (_filter.HasValue("reflections_level")) chatter_reverb_filter.reflectionsLevel = Single.Parse(_filter.GetValue("reflections_level"));
-                    if (_filter.HasValue("reflections_delay")) chatter_reverb_filter.reflectionsDelay = Single.Parse(_filter.GetValue("reflections_delay"));
-                    if (_filter.HasValue("reverb_level")) chatter_reverb_filter.reverbLevel = Single.Parse(_filter.GetValue("reverb_level"));
-                    if (_filter.HasValue("reverb_delay")) chatter_reverb_filter.reverbDelay = Single.Parse(_filter.GetValue("reverb_delay"));
-                    if (_filter.HasValue("diffusion")) chatter_reverb_filter.diffusion = Single.Parse(_filter.GetValue("diffusion"));
-                    if (_filter.HasValue("density")) chatter_reverb_filter.density = Single.Parse(_filter.GetValue("density"));
-                    if (_filter.HasValue("hf_reference")) chatter_reverb_filter.hfReference = Single.Parse(_filter.GetValue("hf_reference"));
-                    if (_filter.HasValue("lf_reference")) chatter_reverb_filter.lFReference = Single.Parse(_filter.GetValue("lf_reference"));
-                }
-            }
-
-            //Beepsources
-            foreach (ConfigNode _source in node.nodes)
-            {
-                if (_source.name == "BEEPSOURCE")
-                {
-                    if (debugging) Debug.Log("[CHATR] loading beepsource");
-                    add_new_beepsource();
-
-                    int x = beepsource_list.Count - 1;
-
-                    if (_source.HasValue("precise")) beepsource_list[x].precise = Boolean.Parse(_source.GetValue("precise"));
-                    if (_source.HasValue("precise_freq"))
-                    {
-                        beepsource_list[x].precise_freq = Int32.Parse(_source.GetValue("precise_freq"));
-                        beepsource_list[x].precise_freq_slider = beepsource_list[x].precise_freq;
-                    }
-                    if (_source.HasValue("loose_freq"))
-                    {
-                        beepsource_list[x].loose_freq = Int32.Parse(_source.GetValue("loose_freq"));
-                        beepsource_list[x].loose_freq_slider = beepsource_list[x].loose_freq;
-                    }
-                    if (_source.HasValue("volume")) beepsource_list[x].audiosource.volume = Single.Parse(_source.GetValue("volume"));
-                    if (_source.HasValue("pitch")) beepsource_list[x].audiosource.pitch = Single.Parse(_source.GetValue("pitch"));
-                    if (_source.HasValue("current_clip")) beepsource_list[x].current_clip = _source.GetValue("current_clip");
-                    if (_source.HasValue("sel_filter")) beepsource_list[x].sel_filter = Int32.Parse(_source.GetValue("sel_filter"));
-                    if (_source.HasValue("show_settings_window")) beepsource_list[x].show_settings_window = Boolean.Parse(_source.GetValue("show_settings_window"));
-                    if (_source.HasValue("reverb_preset_index")) beepsource_list[x].reverb_preset_index = Int32.Parse(_source.GetValue("reverb_preset_index"));
-                    if (_source.HasValue("settings_window_pos_x")) beepsource_list[x].settings_window_pos.x = Single.Parse(_source.GetValue("settings_window_pos_x"));
-                    if (_source.HasValue("settings_window_pos_y")) beepsource_list[x].settings_window_pos.y = Single.Parse(_source.GetValue("settings_window_pos_y"));
-
-                    if (dict_probe_samples.Count > 0)
-                    {
-                        set_beep_clip(beepsource_list[x]);
-
-                        if (beepsource_list[x].precise == false && beepsource_list[x].loose_freq > 0) new_beep_loose_timer_limit(beepsource_list[x]);
-                    }
-
-                    foreach (ConfigNode _filter in _source.nodes)
-                    {
-                        if (_filter.name == "CHORUS")
-                        {
-                            if (_filter.HasValue("enabled"))
-                            {
-                                beepsource_list[x].chorus_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                                //if (debugging) Debug.Log("[CHATR] beepsource index: " + x.ToString() + " :: chorus_enabled = " + beepsource_list[x].chorus_filter.enabled);
-                            }
-                            if (_filter.HasValue("dry_mix")) beepsource_list[x].chorus_filter.dryMix = Single.Parse(_filter.GetValue("dry_mix"));
-                            if (_filter.HasValue("wet_mix_1")) beepsource_list[x].chorus_filter.wetMix1 = Single.Parse(_filter.GetValue("wet_mix_1"));
-                            if (_filter.HasValue("wet_mix_2")) beepsource_list[x].chorus_filter.wetMix2 = Single.Parse(_filter.GetValue("wet_mix_2"));
-                            if (_filter.HasValue("wet_mix_3")) beepsource_list[x].chorus_filter.wetMix3 = Single.Parse(_filter.GetValue("wet_mix_3"));
-                        }
-                        else if (_filter.name == "DISTORTION")
-                        {
-                            if (_filter.HasValue("enabled")) beepsource_list[x].distortion_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                            if (_filter.HasValue("distortion_level")) beepsource_list[x].distortion_filter.distortionLevel = Single.Parse(_filter.GetValue("distortion_level"));
-                        }
-                        else if (_filter.name == "ECHO")
-                        {
-                            if (_filter.HasValue("enabled")) beepsource_list[x].echo_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                            if (_filter.HasValue("delay")) beepsource_list[x].echo_filter.delay = Single.Parse(_filter.GetValue("delay"));
-                            if (_filter.HasValue("decay_ratio")) beepsource_list[x].echo_filter.decayRatio = Single.Parse(_filter.GetValue("decay_ratio"));
-                            if (_filter.HasValue("dry_mix")) beepsource_list[x].echo_filter.dryMix = Single.Parse(_filter.GetValue("dry_mix"));
-                            if (_filter.HasValue("wet_mix")) beepsource_list[x].echo_filter.wetMix = Single.Parse(_filter.GetValue("wet_mix"));
-                        }
-                        else if (_filter.name == "HIGHPASS")
-                        {
-                            if (_filter.HasValue("enabled")) beepsource_list[x].highpass_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                            if (_filter.HasValue("cutoff_freq")) beepsource_list[x].highpass_filter.cutoffFrequency = Single.Parse(_filter.GetValue("cutoff_freq"));
-                            if (_filter.HasValue("resonance_q")) beepsource_list[x].highpass_filter.highpassResonaceQ = Single.Parse(_filter.GetValue("resonance_q"));
-                        }
-                        else if (_filter.name == "LOWPASS")
-                        {
-                            if (_filter.HasValue("enabled")) beepsource_list[x].lowpass_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                            if (_filter.HasValue("cutoff_freq")) beepsource_list[x].lowpass_filter.cutoffFrequency = Single.Parse(_filter.GetValue("cutoff_freq"));
-                            if (_filter.HasValue("resonance_q")) beepsource_list[x].lowpass_filter.lowpassResonaceQ = Single.Parse(_filter.GetValue("resonance_q"));
-                        }
-                        else if (_filter.name == "REVERB")
-                        {
-                            if (_filter.HasValue("enabled")) beepsource_list[x].reverb_filter.enabled = Boolean.Parse(_filter.GetValue("enabled"));
-                            if (_filter.HasValue("reverb_preset")) beepsource_list[x].reverb_filter.reverbPreset = (AudioReverbPreset)Enum.Parse(typeof(AudioReverbPreset), _filter.GetValue("reverb_preset"));
-                            if (_filter.HasValue("dry_level")) beepsource_list[x].reverb_filter.dryLevel = Single.Parse(_filter.GetValue("dry_level"));
-                            if (_filter.HasValue("room")) beepsource_list[x].reverb_filter.room = Single.Parse(_filter.GetValue("room"));
-                            if (_filter.HasValue("room_hf")) beepsource_list[x].reverb_filter.roomHF = Single.Parse(_filter.GetValue("room_hf"));
-                            if (_filter.HasValue("room_lf")) beepsource_list[x].reverb_filter.roomLF = Single.Parse(_filter.GetValue("room_lf"));
-                            if (_filter.HasValue("room_rolloff")) beepsource_list[x].reverb_filter.roomRolloff = Single.Parse(_filter.GetValue("room_rolloff"));
-                            if (_filter.HasValue("decay_time")) beepsource_list[x].reverb_filter.decayTime = Single.Parse(_filter.GetValue("decay_time"));
-                            if (_filter.HasValue("decay_hf_ratio")) beepsource_list[x].reverb_filter.decayHFRatio = Single.Parse(_filter.GetValue("decay_hf_ratio"));
-                            if (_filter.HasValue("reflections_level")) beepsource_list[x].reverb_filter.reflectionsLevel = Single.Parse(_filter.GetValue("reflections_level"));
-                            if (_filter.HasValue("reflections_delay")) beepsource_list[x].reverb_filter.reflectionsDelay = Single.Parse(_filter.GetValue("reflections_delay"));
-                            if (_filter.HasValue("reverb_level")) beepsource_list[x].reverb_filter.reverbLevel = Single.Parse(_filter.GetValue("reverb_level"));
-                            if (_filter.HasValue("reverb_delay")) beepsource_list[x].reverb_filter.reverbDelay = Single.Parse(_filter.GetValue("reverb_delay"));
-                            if (_filter.HasValue("diffusion")) beepsource_list[x].reverb_filter.diffusion = Single.Parse(_filter.GetValue("diffusion"));
-                            if (_filter.HasValue("density")) beepsource_list[x].reverb_filter.density = Single.Parse(_filter.GetValue("density"));
-                            if (_filter.HasValue("hf_reference")) beepsource_list[x].reverb_filter.hfReference = Single.Parse(_filter.GetValue("hf_reference"));
-                            if (_filter.HasValue("lf_reference")) beepsource_list[x].reverb_filter.lFReference = Single.Parse(_filter.GetValue("lf_reference"));
-                        }
-                    }
-                }
-            }
-            if (debugging) Debug.Log("[CHATR] load_shared_settings() END");
-        }
-
+                
         //determine whether the vessel has a part with ModuleRemoteTechSPU and load all relevant RemoteTech variables for the vessel
         public void updateRemoteTechData()
         {
@@ -4478,360 +3095,7 @@ namespace Chatterer
             }
             if (debugging) Debug.Log("[CHATR] single beepsource values pasted from beepsource_clipboard");
         }
-
-        //Functions for per-vessel settings
-        private void new_vessel_node(Vessel v)
-        {
-            //if (debugging) Debug.Log("[CHATR] new_vessel_node() START");
-            ConfigNode vessel_node = new ConfigNode();
-
-            //cn_vessel.name = v.id.ToString();
-            //temp_vessels_node = new ConfigNode();
-            vessel_node.name = "VESSEL";
-
-            vessel_node.AddValue("vessel_name", v.vesselName);
-            vessel_node.AddValue("vessel_id", v.id.ToString());
-
-            save_shared_settings(vessel_node);
-
-
-            /*
-            //
-            vessel_node.AddValue("main_window_pos_x", main_window_pos.x);
-            vessel_node.AddValue("main_window_pos_y", main_window_pos.y);
-            vessel_node.AddValue("ui_icon_pos_x", ui_icon_pos.x);
-            vessel_node.AddValue("ui_icon_pos_y", ui_icon_pos.y);
-            vessel_node.AddValue("main_gui_minimized", main_gui_minimized);
-            vessel_node.AddValue("skin_index", skin_index);
-            vessel_node.AddValue("active_menu", active_menu);
-            vessel_node.AddValue("remotetech_toggle", remotetech_toggle);
-
-            vessel_node.AddValue("chatter_freq", chatter_freq);
-            vessel_node.AddValue("chatter_vol_slider", chatter_vol_slider);
-            vessel_node.AddValue("chatter_sel_filter", chatter_sel_filter);
-            vessel_node.AddValue("show_chatter_filter_settings", show_chatter_filter_settings);
-            vessel_node.AddValue("show_sample_selector", show_sample_selector);
-            vessel_node.AddValue("chatter_reverb_preset_index", chatter_reverb_preset_index);
-            vessel_node.AddValue("chatter_filter_settings_window_pos_x", chatter_filter_settings_window_pos.x);
-            vessel_node.AddValue("chatter_filter_settings_window_pos_y", chatter_filter_settings_window_pos.y);
-
-            vessel_node.AddValue("quindar_toggle", quindar_toggle);
-            vessel_node.AddValue("quindar_vol_slider", quindar_vol_slider);
-            vessel_node.AddValue("sstv_freq", sstv_freq);
-            vessel_node.AddValue("sstv_vol_slider", sstv_vol_slider);
-
-            vessel_node.AddValue("sel_beep_src", sel_beep_src);
-            vessel_node.AddValue("sel_beep_page", sel_beep_page);
-
-
-            //Chatter sets
-            foreach (RBRAudioList chatter_set in chatter_array)
-            {
-                ConfigNode _set = new ConfigNode();
-                _set.name = "AUDIOSET";
-                _set.AddValue("directory", chatter_set.directory);
-                _set.AddValue("is_active", chatter_set.is_active);
-                vessel_node.AddNode(_set);
-            }
-
-            //filters
-            ConfigNode _filter;
-
-            _filter = new ConfigNode();
-            _filter.name = "CHORUS";
-            _filter.AddValue("enabled", chatter_chorus_filter.enabled);
-            _filter.AddValue("dry_mix", chatter_chorus_filter.dryMix);
-            _filter.AddValue("wet_mix_1", chatter_chorus_filter.wetMix1);
-            _filter.AddValue("wet_mix_2", chatter_chorus_filter.wetMix2);
-            _filter.AddValue("wet_mix_3", chatter_chorus_filter.wetMix3);
-            vessel_node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "DISTORTION";
-            _filter.AddValue("enabled", chatter_distortion_filter.enabled);
-            _filter.AddValue("distortion_level", chatter_distortion_filter.distortionLevel);
-            vessel_node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "ECHO";
-            _filter.AddValue("enabled", chatter_echo_filter.enabled);
-            _filter.AddValue("delay", chatter_echo_filter.delay);
-            _filter.AddValue("decay_ratio", chatter_echo_filter.decayRatio);
-            _filter.AddValue("dry_mix", chatter_echo_filter.dryMix);
-            _filter.AddValue("wet_mix", chatter_echo_filter.wetMix);
-            vessel_node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "HIGHPASS";
-            _filter.AddValue("enabled", chatter_highpass_filter.enabled);
-            _filter.AddValue("cutoff_freq", chatter_highpass_filter.cutoffFrequency);
-            _filter.AddValue("resonance_q", chatter_highpass_filter.highpassResonaceQ);
-            vessel_node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "LOWPASS";
-            _filter.AddValue("enabled", chatter_lowpass_filter.enabled);
-            _filter.AddValue("cutoff_freq", chatter_lowpass_filter.cutoffFrequency);
-            _filter.AddValue("resonance_q", chatter_lowpass_filter.lowpassResonaceQ);
-            vessel_node.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "REVERB";
-            _filter.AddValue("enabled", chatter_reverb_filter.enabled);
-            _filter.AddValue("reverb_preset", chatter_reverb_filter.reverbPreset);
-            _filter.AddValue("dry_level", chatter_reverb_filter.dryLevel);
-            _filter.AddValue("room", chatter_reverb_filter.room);
-            _filter.AddValue("room_hf", chatter_reverb_filter.roomHF);
-            _filter.AddValue("room_lf", chatter_reverb_filter.roomLF);
-            _filter.AddValue("room_rolloff", chatter_reverb_filter.roomRolloff);
-            _filter.AddValue("decay_time", chatter_reverb_filter.decayTime);
-            _filter.AddValue("decay_hf_ratio", chatter_reverb_filter.decayHFRatio);
-            _filter.AddValue("reflections_level", chatter_reverb_filter.reflectionsLevel);
-            _filter.AddValue("reflections_delay", chatter_reverb_filter.reflectionsDelay);
-            _filter.AddValue("reverb_level", chatter_reverb_filter.reverbLevel);
-            _filter.AddValue("reverb_delay", chatter_reverb_filter.reverbDelay);
-            _filter.AddValue("diffusion", chatter_reverb_filter.diffusion);
-            _filter.AddValue("density", chatter_reverb_filter.density);
-            _filter.AddValue("hf_reference", chatter_reverb_filter.hfReference);
-            _filter.AddValue("lf_reference", chatter_reverb_filter.lFReference);
-            vessel_node.AddNode(_filter);
-
-
-
-
-            foreach (RBRBeepSource source in beepsource_list)
-            {
-                ConfigNode beep_settings = new ConfigNode();
-                beep_settings.name = "BEEPSOURCE";
-
-                beep_settings.AddValue("precise", source.precise);
-                beep_settings.AddValue("precise_freq", source.precise_freq);
-                beep_settings.AddValue("loose_freq", source.loose_freq);
-                beep_settings.AddValue("volume", source.audiosource.volume);
-                beep_settings.AddValue("pitch", source.audiosource.pitch);
-                beep_settings.AddValue("current_clip", source.current_clip);
-                beep_settings.AddValue("sel_filter", source.sel_filter);
-                beep_settings.AddValue("show_settings_window", source.show_settings_window);
-                beep_settings.AddValue("reverb_preset_index", source.reverb_preset_index);
-                beep_settings.AddValue("settings_window_pos_x", source.settings_window_pos.x);
-                beep_settings.AddValue("settings_window_pos_y", source.settings_window_pos.y);
-
-                //filters
-                //ConfigNode _filter;
-
-                _filter = new ConfigNode();
-                _filter.name = "CHORUS";
-                _filter.AddValue("enabled", source.chorus_filter.enabled);
-                _filter.AddValue("dry_mix", source.chorus_filter.dryMix);
-                _filter.AddValue("wet_mix_1", source.chorus_filter.wetMix1);
-                _filter.AddValue("wet_mix_2", source.chorus_filter.wetMix2);
-                _filter.AddValue("wet_mix_3", source.chorus_filter.wetMix3);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "DISTORTION";
-                _filter.AddValue("enabled", source.distortion_filter.enabled);
-                _filter.AddValue("distortion_level", source.distortion_filter.distortionLevel);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "ECHO";
-                _filter.AddValue("enabled", source.echo_filter.enabled);
-                _filter.AddValue("delay", source.echo_filter.delay);
-                _filter.AddValue("decay_ratio", source.echo_filter.decayRatio);
-                _filter.AddValue("dry_mix", source.echo_filter.dryMix);
-                _filter.AddValue("wet_mix", source.echo_filter.wetMix);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "HIGHPASS";
-                _filter.AddValue("enabled", source.highpass_filter.enabled);
-                _filter.AddValue("cutoff_freq", source.highpass_filter.cutoffFrequency);
-                _filter.AddValue("resonance_q", source.highpass_filter.highpassResonaceQ);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "LOWPASS";
-                _filter.AddValue("enabled", source.lowpass_filter.enabled);
-                _filter.AddValue("cutoff_freq", source.lowpass_filter.cutoffFrequency);
-                _filter.AddValue("resonance_q", source.lowpass_filter.lowpassResonaceQ);
-                beep_settings.AddNode(_filter);
-
-                _filter = new ConfigNode();
-                _filter.name = "REVERB";
-                _filter.AddValue("enabled", source.reverb_filter.enabled);
-                _filter.AddValue("reverb_preset", source.reverb_filter.reverbPreset);
-                _filter.AddValue("dry_level", source.reverb_filter.dryLevel);
-                _filter.AddValue("room", source.reverb_filter.room);
-                _filter.AddValue("room_hf", source.reverb_filter.roomHF);
-                _filter.AddValue("room_lf", source.reverb_filter.roomLF);
-                _filter.AddValue("room_rolloff", source.reverb_filter.roomRolloff);
-                _filter.AddValue("decay_time", source.reverb_filter.decayTime);
-                _filter.AddValue("decay_hf_ratio", source.reverb_filter.decayHFRatio);
-                _filter.AddValue("reflections_level", source.reverb_filter.reflectionsLevel);
-                _filter.AddValue("reflections_delay", source.reverb_filter.reflectionsDelay);
-                _filter.AddValue("reverb_level", source.reverb_filter.reverbLevel);
-                _filter.AddValue("reverb_delay", source.reverb_filter.reverbDelay);
-                _filter.AddValue("diffusion", source.reverb_filter.diffusion);
-                _filter.AddValue("density", source.reverb_filter.density);
-                _filter.AddValue("hf_reference", source.reverb_filter.hfReference);
-                _filter.AddValue("lf_reference", source.reverb_filter.lFReference);
-                beep_settings.AddNode(_filter);
-
-                vessel_node.AddNode(beep_settings);
-            }
-            */
-            vessel_settings_node.AddNode(vessel_node);
-            vessel_settings_node.Save(settings_path + "vessels.cfg");
-
-
-            //if (debugging) Debug.Log("[CHATR] new_vessel_node() :: vessel_node added to vessel_settings_node");
-        }
-
-        private void load_vessel_settings_node()
-        {
-            //if (debugging) Debug.Log("[CHATR] START load_vessel_settings_node()");
-            vessel_settings_node = ConfigNode.Load(settings_path + "vessels.cfg");
-
-            if (vessel_settings_node != null)
-            {
-                if (debugging) Debug.Log("[CHATR] load_vessel_settings_node() :: vessel_settings.cfg loaded OK");
-                //now search for a matching vessel_id
-                //search_vessel_settings_node();
-            }
-            else
-            {
-                if (debugging) Debug.LogWarning("[CHATR] load_vessel_settings_node() :: vessel_settings.cfg is null, creating a new one");
-                vessel_settings_node = new ConfigNode();
-                vessel_settings_node.name = "FLIGHTS";
-                new_vessel_node(vessel);  //add current vessel to vessel_settings_node
-                //save_vessel_settings_node();
-                //if (debugging) Debug.Log("[CHATR] load_vessel_settings_node() :: current vessel node saved to vessel_settings.cfg");
-            }
-
-        }
-
-        private void load_vessel_node(ConfigNode node)
-        {
-            if (debugging) Debug.Log("[CHATR] load_vessel_node() :: loading vessel settings for this vessel from node");
-
-            //destroy_all_beep_players();
-            //destroy_all_background_players();
-
-            load_shared_settings(node);
-
-            if (chatter_array.Count == 0)
-            {
-                if (debugging) Debug.Log("[CHATR] No audiosets found in config, adding defaults");
-                add_default_audiosets();
-            }
-
-            if (beepsource_list.Count == 0)
-            {
-                if (debugging) Debug.LogWarning("[CHATR] beepsource_list.Count == 0, adding default 3");
-                add_default_beepsources();
-            }
-
-            if (backgroundsource_list.Count == 0)
-            {
-                if (debugging) Debug.LogWarning("[CHATR] backgroundsource_list.Count == 0, adding default 2");
-                add_default_backgroundsources();
-            }
-
-            if (debugging) Debug.Log("[CHATR] load_vessel_node() :: vessel settings loaded OK : total beep sources = " + beepsource_list.Count);
-        }
-
-        private void search_vessel_settings_node()
-        {
-            if (debugging) Debug.Log("[CHATR] START search_vessel_settings_node()");
-
-            bool no_match = true;
-
-            if (debugging) Debug.Log("[CHATR] active vessel id = " + vessel.id.ToString());
-
-            foreach (ConfigNode n in vessel_settings_node.nodes)
-            {
-                string val = n.GetValue("vessel_id");
-                if (debugging) Debug.Log("[CHATR] n.GetValue(\"vessel_id\") = " + n.GetValue("vessel_id"));
-                if (val == vessel.id.ToString())
-                {
-                    if (debugging) Debug.Log("[CHATR] search_vessel_settings_node() :: vessel_id match");
-                    load_vessel_node(n);    //load vals
-                    no_match = false;
-                    break;
-                }
-                else if (debugging) Debug.Log("[CHATR] no match, continuing search...");
-            }
-            if (no_match)
-            {
-                if (debugging) Debug.Log("[CHATR] finished search, no vessel_id match :: creating new node for this vessel");
-                //temp_vessels_node = new ConfigNode();
-                new_vessel_node(vessel);
-                //save_vessel_settings_node();  //done in new_vessel_node
-                if (debugging) Debug.Log("[CHATR] new vessel node created and saved");
-                load_chatter_audio();   //load audio in case there is none
-                //return;
-            }
-
-            //save_vessel_settings_node();
-        }
-
-        private void write_vessel_settings()
-        {
-            //update vessel_settings.cfg here also
-            //get all nodes from vessel_settings_node that are not the active vessel and put them into a list all_but_curr
-            //set vessel_settings_node = all_but_curr
-            //create a new node for the active vessel with its current settings
-            //add new node to vessel_settings_node
-            //save vessel_settings_node to .cfg
-            if (debugging) Debug.Log("[CHATR] writing vessel_settings node to disk");
-
-
-
-            ConfigNode all_but_curr_vessel = new ConfigNode();
-
-            if (debugging) Debug.Log("[CHATR] active vessel.id = " + vessel.id.ToString());
-            foreach (ConfigNode cn in vessel_settings_node.nodes)
-            {
-                //
-                if (cn.HasValue("vessel_id"))
-                {
-                    string val = cn.GetValue("vessel_id");
-                    if (debugging) Debug.Log("[CHATR] node vessel_id = " + val);
-
-                    if (val != vessel.id.ToString())
-                    {
-                        //found an id that is not the current vessel
-                        //add it to the list
-
-                        all_but_curr_vessel.AddNode(cn);
-                        if (debugging) Debug.Log("[CHATR] write_vessel_settings() :: node vessel_id != vessel.id :: node vessel added to all_but_curr_vessel");
-                    }
-                    //else
-                    //{
-                    //    all_but_prev_vessel.AddNode(cn);
-                    //}
-                }
-            }
-            //foreach (ConfigNode cn in vessel_settings_node.nodes)
-            //{
-            //vessel_settings_node.RemoveNodes("");
-            //    if (debugging) Debug.Log("[CHATR] old nodes removed");
-            //}
-
-            vessel_settings_node = all_but_curr_vessel;
-            //if (debugging) Debug.Log("[CHATR] write_vessel_settings() :: vessel_settings node = all_but_curr_vessel");
-
-            new_vessel_node(vessel);
-            //if (debugging) Debug.Log("[CHATR] write_vessel_settings() :: new node created using vessel and added to vessel_settings node");
-
-            //save_vessel_settings_node();
-            vessel_settings_node.Save(settings_path + "vessels.cfg");
-            if (debugging) Debug.Log("[CHATR] write_vessel_settings() END :: vessel_settings node saved to vessel_settings.cfg");
-            //end func
-            if (debugging) Debug.Log("[CHATR] vessel_settings node saved to disk :: vessel node count = " + vessel_settings_node.nodes.Count);
-        }
-
+        
         //Set some default stuff
         private void add_default_audiosets()
         {
@@ -4995,365 +3259,7 @@ namespace Chatterer
                 GUI.Box(new Rect(x, Event.current.mousePosition.y, w, 25f), GUI.tooltip, gs_tooltip);
             }
         }
-
-        //Reset filter
-        private void reset_chorus_filter(AudioChorusFilter acf)
-        {
-            //reset chorus filter to default
-            foreach (ConfigNode filter in filter_defaults.nodes)
-            {
-                if (filter.name == "CHORUS")
-                {
-                    if (filter.HasValue("dry_mix")) acf.dryMix = Single.Parse(filter.GetValue("dry_mix"));
-                    if (filter.HasValue("wet_mix_1")) acf.wetMix1 = Single.Parse(filter.GetValue("wet_mix_1"));
-                    if (filter.HasValue("wet_mix_2")) acf.wetMix2 = Single.Parse(filter.GetValue("wet_mix_2"));
-                    if (filter.HasValue("wet_mix_3")) acf.wetMix3 = Single.Parse(filter.GetValue("wet_mix_3"));
-                    if (filter.HasValue("delay")) acf.delay = Single.Parse(filter.GetValue("delay"));
-                    if (filter.HasValue("rate")) acf.rate = Single.Parse(filter.GetValue("rate"));
-                    if (filter.HasValue("depth")) acf.depth = Single.Parse(filter.GetValue("depth"));
-                }
-            }
-        }
-
-        private void reset_distortion_filter(AudioDistortionFilter adf)
-        {
-            //reset distortion filter to default
-            foreach (ConfigNode filter in filter_defaults.nodes)
-            {
-                if (filter.name == "DISTORTION")
-                {
-                    if (filter.HasValue("distortion_level")) adf.distortionLevel = Single.Parse(filter.GetValue("distortion_level"));
-                }
-            }
-        }
-
-        private void reset_echo_filter(AudioEchoFilter aef)
-        {
-            //reset echo filter to default
-            foreach (ConfigNode filter in filter_defaults.nodes)
-            {
-                if (filter.name == "ECHO")
-                {
-                    if (filter.HasValue("delay")) aef.delay = Single.Parse(filter.GetValue("delay"));
-                    if (filter.HasValue("decay_ratio")) aef.decayRatio = Single.Parse(filter.GetValue("decay_ratio"));
-                    if (filter.HasValue("dry_mix")) aef.dryMix = Single.Parse(filter.GetValue("dry_mix"));
-                    if (filter.HasValue("wet_mix")) aef.wetMix = Single.Parse(filter.GetValue("wet_mix"));
-                }
-            }
-        }
-
-        private void reset_highpass_filter(AudioHighPassFilter ahpf)
-        {
-            //reset highpass filter to default
-            foreach (ConfigNode filter in filter_defaults.nodes)
-            {
-                if (filter.name == "HIGHPASS")
-                {
-                    if (filter.HasValue("cutoff_freq")) ahpf.cutoffFrequency = Single.Parse(filter.GetValue("cutoff_freq"));
-                    if (filter.HasValue("resonance_q")) ahpf.highpassResonaceQ = Single.Parse(filter.GetValue("resonance_q"));
-                }
-            }
-        }
-
-        private void reset_lowpass_filter(AudioLowPassFilter alpf)
-        {
-            //reset lowpass filter to default
-            foreach (ConfigNode filter in filter_defaults.nodes)
-            {
-                if (filter.name == "LOWPASS")
-                {
-                    if (filter.HasValue("cutoff_freq")) alpf.cutoffFrequency = Single.Parse(filter.GetValue("cutoff_freq"));
-                    if (filter.HasValue("resonance_q")) alpf.lowpassResonaceQ = Single.Parse(filter.GetValue("resonance_q"));
-                }
-            }
-        }
-
-        private void reset_reverb_filter(AudioReverbFilter arf)
-        {
-            //reset reverb filter to default
-            foreach (ConfigNode filter in filter_defaults.nodes)
-            {
-                if (filter.name == "REVERB")
-                {
-                    if (filter.HasValue("reverb_preset")) arf.reverbPreset = (AudioReverbPreset)Enum.Parse(typeof(AudioReverbPreset), filter.GetValue("reverb_preset"));
-                    if (filter.HasValue("dry_level")) arf.dryLevel = Single.Parse(filter.GetValue("dry_level"));
-                    if (filter.HasValue("room")) arf.room = Single.Parse(filter.GetValue("room"));
-                    if (filter.HasValue("room_hf")) arf.roomHF = Single.Parse(filter.GetValue("room_hf"));
-                    if (filter.HasValue("room_lf")) arf.roomLF = Single.Parse(filter.GetValue("room_lf"));
-                    if (filter.HasValue("room_rolloff")) arf.roomRolloff = Single.Parse(filter.GetValue("room_rolloff"));
-                    if (filter.HasValue("decay_time")) arf.decayTime = Single.Parse(filter.GetValue("decay_time"));
-                    if (filter.HasValue("decay_hf_ratio")) arf.decayHFRatio = Single.Parse(filter.GetValue("decay_hf_ratio"));
-                    if (filter.HasValue("reflections_level")) arf.reflectionsLevel = Single.Parse(filter.GetValue("reflections_level"));
-                    if (filter.HasValue("reflections_delay")) arf.reflectionsDelay = Single.Parse(filter.GetValue("reflections_delay"));
-                    if (filter.HasValue("reverb_level")) arf.reverbLevel = Single.Parse(filter.GetValue("reverb_level"));
-                    if (filter.HasValue("reverb_delay")) arf.reverbDelay = Single.Parse(filter.GetValue("reverb_delay"));
-                    if (filter.HasValue("diffusion")) arf.diffusion = Single.Parse(filter.GetValue("diffusion"));
-                    if (filter.HasValue("density")) arf.density = Single.Parse(filter.GetValue("density"));
-                    if (filter.HasValue("hf_reference")) arf.hfReference = Single.Parse(filter.GetValue("hf_reference"));
-                    if (filter.HasValue("lf_reference")) arf.lFReference = Single.Parse(filter.GetValue("lf_reference"));
-                }
-            }
-        }
-
-        //Copy/Paste chatter filters
-        private void copy_all_chatter_filters()
-        {
-            //copy all chatter filter settings to a temp clipboard ConfigNode
-
-            filters_clipboard = new ConfigNode();
-
-            ConfigNode _filter;
-
-            _filter = new ConfigNode();
-            _filter.name = "CHORUS";
-            _filter.AddValue("enabled", chatter_chorus_filter.enabled);
-            _filter.AddValue("dry_mix", chatter_chorus_filter.dryMix);
-            _filter.AddValue("wet_mix_1", chatter_chorus_filter.wetMix1);
-            _filter.AddValue("wet_mix_2", chatter_chorus_filter.wetMix2);
-            _filter.AddValue("wet_mix_3", chatter_chorus_filter.wetMix3);
-            _filter.AddValue("delay", chatter_chorus_filter.delay);
-            _filter.AddValue("rate", chatter_chorus_filter.rate);
-            _filter.AddValue("depth", chatter_chorus_filter.depth);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "DISTORTION";
-            _filter.AddValue("enabled", chatter_distortion_filter.enabled);
-            _filter.AddValue("distortion_level", chatter_distortion_filter.distortionLevel);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "ECHO";
-            _filter.AddValue("enabled", chatter_echo_filter.enabled);
-            _filter.AddValue("delay", chatter_echo_filter.delay);
-            _filter.AddValue("decay_ratio", chatter_echo_filter.decayRatio);
-            _filter.AddValue("dry_mix", chatter_echo_filter.dryMix);
-            _filter.AddValue("wet_mix", chatter_echo_filter.wetMix);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "HIGHPASS";
-            _filter.AddValue("enabled", chatter_highpass_filter.enabled);
-            _filter.AddValue("cutoff_freq", chatter_highpass_filter.cutoffFrequency);
-            _filter.AddValue("resonance_q", chatter_highpass_filter.highpassResonaceQ);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "LOWPASS";
-            _filter.AddValue("enabled", chatter_lowpass_filter.enabled);
-            _filter.AddValue("cutoff_freq", chatter_lowpass_filter.cutoffFrequency);
-            _filter.AddValue("resonance_q", chatter_lowpass_filter.lowpassResonaceQ);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "REVERB";
-            _filter.AddValue("enabled", chatter_reverb_filter.enabled);
-            _filter.AddValue("reverb_preset", chatter_reverb_filter.reverbPreset);
-            _filter.AddValue("dry_level", chatter_reverb_filter.dryLevel);
-            _filter.AddValue("room", chatter_reverb_filter.room);
-            _filter.AddValue("room_hf", chatter_reverb_filter.roomHF);
-            _filter.AddValue("room_lf", chatter_reverb_filter.roomLF);
-            _filter.AddValue("room_rolloff", chatter_reverb_filter.roomRolloff);
-            _filter.AddValue("decay_time", chatter_reverb_filter.decayTime);
-            _filter.AddValue("decay_hf_ratio", chatter_reverb_filter.decayHFRatio);
-            _filter.AddValue("reflections_level", chatter_reverb_filter.reflectionsLevel);
-            _filter.AddValue("reflections_delay", chatter_reverb_filter.reflectionsDelay);
-            _filter.AddValue("reverb_level", chatter_reverb_filter.reverbLevel);
-            _filter.AddValue("reverb_delay", chatter_reverb_filter.reverbDelay);
-            _filter.AddValue("diffusion", chatter_reverb_filter.diffusion);
-            _filter.AddValue("density", chatter_reverb_filter.density);
-            _filter.AddValue("hf_reference", chatter_reverb_filter.hfReference);
-            _filter.AddValue("lf_reference", chatter_reverb_filter.lFReference);
-            filters_clipboard.AddNode(_filter);
-
-            if (debugging) Debug.Log("[CHATR] all chatter filter values copied to filters_clipboard");
-        }
-
-        private void paste_all_chatter_filters()
-        {
-            ConfigNode filter = new ConfigNode();
-
-            filter = filters_clipboard.GetNode("CHORUS");
-            if (filter.HasValue("enabled")) chatter_chorus_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("dry_mix")) chatter_chorus_filter.dryMix = Single.Parse(filter.GetValue("dry_mix"));
-            if (filter.HasValue("wet_mix_1")) chatter_chorus_filter.wetMix1 = Single.Parse(filter.GetValue("wet_mix_1"));
-            if (filter.HasValue("wet_mix_2")) chatter_chorus_filter.wetMix2 = Single.Parse(filter.GetValue("wet_mix_2"));
-            if (filter.HasValue("wet_mix_3")) chatter_chorus_filter.wetMix3 = Single.Parse(filter.GetValue("wet_mix_3"));
-            if (filter.HasValue("delay")) chatter_chorus_filter.delay = Single.Parse(filter.GetValue("delay"));
-            if (filter.HasValue("rate")) chatter_chorus_filter.rate = Single.Parse(filter.GetValue("rate"));
-            if (filter.HasValue("depth")) chatter_chorus_filter.depth = Single.Parse(filter.GetValue("depth"));
-
-            filter = filters_clipboard.GetNode("DISTORTION");
-            if (filter.HasValue("enabled")) chatter_distortion_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("distortion_level")) chatter_distortion_filter.distortionLevel = Single.Parse(filter.GetValue("distortion_level"));
-
-            filter = filters_clipboard.GetNode("ECHO");
-            if (filter.HasValue("enabled")) chatter_echo_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("delay")) chatter_echo_filter.delay = Single.Parse(filter.GetValue("delay"));
-            if (filter.HasValue("decay_ratio")) chatter_echo_filter.decayRatio = Single.Parse(filter.GetValue("decay_ratio"));
-            if (filter.HasValue("dry_mix")) chatter_echo_filter.dryMix = Single.Parse(filter.GetValue("dry_mix"));
-            if (filter.HasValue("wet_mix")) chatter_echo_filter.wetMix = Single.Parse(filter.GetValue("wet_mix"));
-
-            filter = filters_clipboard.GetNode("HIGHPASS");
-            if (filter.HasValue("enabled")) chatter_highpass_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("cutoff_freq")) chatter_highpass_filter.cutoffFrequency = Single.Parse(filter.GetValue("cutoff_freq"));
-            if (filter.HasValue("resonance_q")) chatter_highpass_filter.highpassResonaceQ = Single.Parse(filter.GetValue("resonance_q"));
-
-            filter = filters_clipboard.GetNode("LOWPASS");
-            if (filter.HasValue("enabled")) chatter_lowpass_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("cutoff_freq")) chatter_lowpass_filter.cutoffFrequency = Single.Parse(filter.GetValue("cutoff_freq"));
-            if (filter.HasValue("resonance_q")) chatter_lowpass_filter.lowpassResonaceQ = Single.Parse(filter.GetValue("resonance_q"));
-
-            filter = filters_clipboard.GetNode("REVERB");
-            if (filter.HasValue("enabled")) chatter_reverb_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("reverb_preset")) chatter_reverb_filter.reverbPreset = (AudioReverbPreset)Enum.Parse(typeof(AudioReverbPreset), filter.GetValue("reverb_preset"));
-            if (filter.HasValue("dry_level")) chatter_reverb_filter.dryLevel = Single.Parse(filter.GetValue("dry_level"));
-            if (filter.HasValue("room")) chatter_reverb_filter.room = Single.Parse(filter.GetValue("room"));
-            if (filter.HasValue("room_hf")) chatter_reverb_filter.roomHF = Single.Parse(filter.GetValue("room_hf"));
-            if (filter.HasValue("room_lf")) chatter_reverb_filter.roomLF = Single.Parse(filter.GetValue("room_lf"));
-            if (filter.HasValue("room_rolloff")) chatter_reverb_filter.roomRolloff = Single.Parse(filter.GetValue("room_rolloff"));
-            if (filter.HasValue("decay_time")) chatter_reverb_filter.decayTime = Single.Parse(filter.GetValue("decay_time"));
-            if (filter.HasValue("decay_hf_ratio")) chatter_reverb_filter.decayHFRatio = Single.Parse(filter.GetValue("decay_hf_ratio"));
-            if (filter.HasValue("reflections_level")) chatter_reverb_filter.reflectionsLevel = Single.Parse(filter.GetValue("reflections_level"));
-            if (filter.HasValue("reflections_delay")) chatter_reverb_filter.reflectionsDelay = Single.Parse(filter.GetValue("reflections_delay"));
-            if (filter.HasValue("reverb_level")) chatter_reverb_filter.reverbLevel = Single.Parse(filter.GetValue("reverb_level"));
-            if (filter.HasValue("reverb_delay")) chatter_reverb_filter.reverbDelay = Single.Parse(filter.GetValue("reverb_delay"));
-            if (filter.HasValue("diffusion")) chatter_reverb_filter.diffusion = Single.Parse(filter.GetValue("diffusion"));
-            if (filter.HasValue("density")) chatter_reverb_filter.density = Single.Parse(filter.GetValue("density"));
-            if (filter.HasValue("hf_reference")) chatter_reverb_filter.hfReference = Single.Parse(filter.GetValue("hf_reference"));
-            if (filter.HasValue("lf_reference")) chatter_reverb_filter.lFReference = Single.Parse(filter.GetValue("lf_reference"));
-
-            if (debugging) Debug.Log("[CHATR] all chatter filter values pasted from filters_clipboard");
-        }
-
-        //Copy/Paste beep filters
-        private void copy_all_beep_filters(BeepSource source)
-        {
-            filters_clipboard = new ConfigNode();
-            ConfigNode _filter;
-
-            _filter = new ConfigNode();
-            _filter.name = "CHORUS";
-            _filter.AddValue("enabled", source.chorus_filter.enabled);
-            _filter.AddValue("dry_mix", source.chorus_filter.dryMix);
-            _filter.AddValue("wet_mix_1", source.chorus_filter.wetMix1);
-            _filter.AddValue("wet_mix_2", source.chorus_filter.wetMix2);
-            _filter.AddValue("wet_mix_3", source.chorus_filter.wetMix3);
-            _filter.AddValue("delay", source.chorus_filter.delay);
-            _filter.AddValue("rate", source.chorus_filter.rate);
-            _filter.AddValue("depth", source.chorus_filter.depth);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "DISTORTION";
-            _filter.AddValue("enabled", source.distortion_filter.enabled);
-            _filter.AddValue("distortion_level", source.distortion_filter.distortionLevel);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "ECHO";
-            _filter.AddValue("enabled", source.echo_filter.enabled);
-            _filter.AddValue("delay", source.echo_filter.delay);
-            _filter.AddValue("decay_ratio", source.echo_filter.decayRatio);
-            _filter.AddValue("dry_mix", source.echo_filter.dryMix);
-            _filter.AddValue("wet_mix", source.echo_filter.wetMix);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "HIGHPASS";
-            _filter.AddValue("enabled", source.highpass_filter.enabled);
-            _filter.AddValue("cutoff_freq", source.highpass_filter.cutoffFrequency);
-            _filter.AddValue("resonance_q", source.highpass_filter.highpassResonaceQ);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "LOWPASS";
-            _filter.AddValue("enabled", source.lowpass_filter.enabled);
-            _filter.AddValue("cutoff_freq", source.lowpass_filter.cutoffFrequency);
-            _filter.AddValue("resonance_q", source.lowpass_filter.lowpassResonaceQ);
-            filters_clipboard.AddNode(_filter);
-
-            _filter = new ConfigNode();
-            _filter.name = "REVERB";
-            _filter.AddValue("enabled", source.reverb_filter.enabled);
-            _filter.AddValue("reverb_preset", source.reverb_filter.reverbPreset);
-            _filter.AddValue("dry_level", source.reverb_filter.dryLevel);
-            _filter.AddValue("room", source.reverb_filter.room);
-            _filter.AddValue("room_hf", source.reverb_filter.roomHF);
-            _filter.AddValue("room_lf", source.reverb_filter.roomLF);
-            _filter.AddValue("room_rolloff", source.reverb_filter.roomRolloff);
-            _filter.AddValue("decay_time", source.reverb_filter.decayTime);
-            _filter.AddValue("decay_hf_ratio", source.reverb_filter.decayHFRatio);
-            _filter.AddValue("reflections_level", source.reverb_filter.reflectionsLevel);
-            _filter.AddValue("reflections_delay", source.reverb_filter.reflectionsDelay);
-            _filter.AddValue("reverb_level", source.reverb_filter.reverbLevel);
-            _filter.AddValue("reverb_delay", source.reverb_filter.reverbDelay);
-            _filter.AddValue("diffusion", source.reverb_filter.diffusion);
-            _filter.AddValue("density", source.reverb_filter.density);
-            _filter.AddValue("hf_reference", source.reverb_filter.hfReference);
-            _filter.AddValue("lf_reference", source.reverb_filter.lFReference);
-            filters_clipboard.AddNode(_filter);
-
-            if (debugging) Debug.Log("[CHATR] all beep filter values copied to filters_clipboard");
-        }
-
-        private void paste_all_beep_filters(BeepSource source)
-        {
-            ConfigNode filter = new ConfigNode();
-
-            filter = filters_clipboard.GetNode("CHORUS");
-            if (filter.HasValue("enabled")) source.chorus_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("dry_mix")) source.chorus_filter.dryMix = Single.Parse(filter.GetValue("dry_mix"));
-            if (filter.HasValue("wet_mix_1")) source.chorus_filter.wetMix1 = Single.Parse(filter.GetValue("wet_mix_1"));
-            if (filter.HasValue("wet_mix_2")) source.chorus_filter.wetMix2 = Single.Parse(filter.GetValue("wet_mix_2"));
-            if (filter.HasValue("wet_mix_3")) source.chorus_filter.wetMix3 = Single.Parse(filter.GetValue("wet_mix_3"));
-            if (filter.HasValue("delay")) source.chorus_filter.dryMix = Single.Parse(filter.GetValue("delay"));
-            if (filter.HasValue("rate")) source.chorus_filter.dryMix = Single.Parse(filter.GetValue("rate"));
-            if (filter.HasValue("depth")) source.chorus_filter.dryMix = Single.Parse(filter.GetValue("depth"));
-
-            filter = filters_clipboard.GetNode("DISTORTION");
-            if (filter.HasValue("enabled")) source.distortion_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("distortion_level")) source.distortion_filter.distortionLevel = Single.Parse(filter.GetValue("distortion_level"));
-
-            filter = filters_clipboard.GetNode("ECHO");
-            if (filter.HasValue("enabled")) source.echo_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("delay")) source.echo_filter.delay = Single.Parse(filter.GetValue("delay"));
-            if (filter.HasValue("decay_ratio")) source.echo_filter.decayRatio = Single.Parse(filter.GetValue("decay_ratio"));
-            if (filter.HasValue("dry_mix")) source.echo_filter.dryMix = Single.Parse(filter.GetValue("dry_mix"));
-            if (filter.HasValue("wet_mix")) source.echo_filter.wetMix = Single.Parse(filter.GetValue("wet_mix"));
-
-            filter = filters_clipboard.GetNode("HIGHPASS");
-            if (filter.HasValue("enabled")) source.highpass_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("cutoff_freq")) source.highpass_filter.cutoffFrequency = Single.Parse(filter.GetValue("cutoff_freq"));
-            if (filter.HasValue("resonance_q")) source.highpass_filter.highpassResonaceQ = Single.Parse(filter.GetValue("resonance_q"));
-
-            filter = filters_clipboard.GetNode("LOWPASS");
-            if (filter.HasValue("enabled")) source.lowpass_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("cutoff_freq")) source.lowpass_filter.cutoffFrequency = Single.Parse(filter.GetValue("cutoff_freq"));
-            if (filter.HasValue("resonance_q")) source.lowpass_filter.lowpassResonaceQ = Single.Parse(filter.GetValue("resonance_q"));
-
-            filter = filters_clipboard.GetNode("REVERB");
-            if (filter.HasValue("enabled")) source.reverb_filter.enabled = Boolean.Parse(filter.GetValue("enabled"));
-            if (filter.HasValue("reverb_preset")) source.reverb_filter.reverbPreset = (AudioReverbPreset)Enum.Parse(typeof(AudioReverbPreset), filter.GetValue("reverb_preset"));
-            if (filter.HasValue("dry_level")) source.reverb_filter.dryLevel = Single.Parse(filter.GetValue("dry_level"));
-            if (filter.HasValue("room")) source.reverb_filter.room = Single.Parse(filter.GetValue("room"));
-            if (filter.HasValue("room_hf")) source.reverb_filter.roomHF = Single.Parse(filter.GetValue("room_hf"));
-            if (filter.HasValue("room_lf")) source.reverb_filter.roomLF = Single.Parse(filter.GetValue("room_lf"));
-            if (filter.HasValue("room_rolloff")) source.reverb_filter.roomRolloff = Single.Parse(filter.GetValue("room_rolloff"));
-            if (filter.HasValue("decay_time")) source.reverb_filter.decayTime = Single.Parse(filter.GetValue("decay_time"));
-            if (filter.HasValue("decay_hf_ratio")) source.reverb_filter.decayHFRatio = Single.Parse(filter.GetValue("decay_hf_ratio"));
-            if (filter.HasValue("reflections_level")) source.reverb_filter.reflectionsLevel = Single.Parse(filter.GetValue("reflections_level"));
-            if (filter.HasValue("reflections_delay")) source.reverb_filter.reflectionsDelay = Single.Parse(filter.GetValue("reflections_delay"));
-            if (filter.HasValue("reverb_level")) source.reverb_filter.reverbLevel = Single.Parse(filter.GetValue("reverb_level"));
-            if (filter.HasValue("reverb_delay")) source.reverb_filter.reverbDelay = Single.Parse(filter.GetValue("reverb_delay"));
-            if (filter.HasValue("diffusion")) source.reverb_filter.diffusion = Single.Parse(filter.GetValue("diffusion"));
-            if (filter.HasValue("density")) source.reverb_filter.density = Single.Parse(filter.GetValue("density"));
-            if (filter.HasValue("hf_reference")) source.reverb_filter.hfReference = Single.Parse(filter.GetValue("hf_reference"));
-            if (filter.HasValue("lf_reference")) source.reverb_filter.lFReference = Single.Parse(filter.GetValue("lf_reference"));
-
-            if (debugging) Debug.Log("[CHATR] all beep filter values pasted from filters_clipboard");
-        }
-
+        
         //Main
         public void Awake()
         {
